### 2023-10-02 4564e30
Runnable code with partial doc strings.
