### 2023-10-02 0728379
remove files for 2018 paper
The JOSS buildbot is building the wrong paper, so temporarily remove the original manuscript files in this git branch so that there is only one set of files (the 2023 submission).
