+++ paper/2018/paper.bib
@@ -1,110 +1,0 @@
-@inproceedings{Andrews2014,
-	Author = {R. W. Andrews and J. S. Stein and C. Hansen and D. Riley},
-	Booktitle = {2014 IEEE 40th Photovoltaic Specialist Conference (PVSC)},
-	doi = {10.1109/PVSC.2014.6925501},
-	Title = {Introduction to the open source PV LIB for python Photovoltaic system modelling package},
-	Year = {2014}}
-
-@online{DOESF2,
-    author = {Department of Energy},
-    title = {Solar Forecasting 2},
-    year = 2018,
-    url = {https://www.energy.gov/eere/solar/solar-forecasting-2},
-    urldate = {2018-08-02}}
-
-@article{Gagne2017,
-	Author = {Gagne II, David John and McGovern, Amy and Haupt, Sue Ellen and Williams, John K.},
-	Journal = {Solar Energy},
-	Pages = {383- 393},
-	doi = {10.1016/j.solener.2017.04.031},
-	Title = {Evaluation of statistical learning configurations for gridded solar irradiance forecasting},
-	Volume = {150},
-	Year = {2017}}
-
-@inproceedings{Holmgren2015,
-	Author = {W. F. Holmgren and R. W. Andrews and A. T. Lorenzo and J. S. Stein},
-	Booktitle = {2015 IEEE 42nd Photovoltaic Specialist Conference (PVSC)},
-	Month = {June},
-	Pages = {1-5},
-	Title = {PVLIB Python 2015},
-	Year = {2015},
-	doi = {10.1109/PVSC.2015.7356005}}
-
-@inproceedings{Holmgren2016,
-	Author = {W. F. Holmgren and D. G. Groenendyk},
-	Booktitle = {2016 IEEE 43rd Photovoltaic Specialists Conference (PVSC)},
-	Month = {June},
-	Pages = {0972-0975},
-	Title = {An open source solar power forecasting tool using PVLIB-Python},
-	Year = {2016},
-	doi = {10.1109/PVSC.2016.7749755}}
-
-@inproceedings{Holmgren2017,
-	Author = {William F. Holmgren and Antonio T. Lorenzo and Clifford Hansen},
-	Booktitle = {2017 IEEE 44th Photovoltaic Specialists Conference},
-	Title = {A Comparison of PV Power Forecasts Using PVLib-Python},
-	Year = {2017},
-	doi = {10.5281/zenodo.1400857}}
-
-@inproceedings{Holmgren2018,
-	Author = {William F. Holmgren and Clifford W. Hansen and Joshua S. Stein and Mark A. Mikofski},
-	Booktitle = {2018 IEEE 45th Photovoltaic Specialists Conference},
-	Title = {Review of open source tools for PV modeling},
-	Year = {2018},
-	doi = {10.5281/zenodo.1401378}}
-
-@article{Louwen2017,
-	Author = {Atse Louwen and Ruud E.I. Schropp and Wilfried G.J.H.M. van Sark and Andr{\'e} P.C. Faaij},
-	Journal = {Solar Energy},
-	Pages = {1339 - 1353},
-	Title = {Geospatial analysis of the energy yield and environmental footprint of different photovoltaic module technologies},
-	Volume = {155},
-	Year = {2017},
-	doi = {10.1016/j.solener.2017.07.056}}
-
-@inproceedings{Mikofski2016,
-	Author = {M. Mikofski and A. Oumbe and C. Li and B. Bourne},
-	Booktitle = {2016 IEEE 43rd Photovoltaic Specialists Conference (PVSC)},
-	Month = {June},
-	Pages = {1357-1362},
[…]
