### 2023-10-02 #5910 @kandersolar
@rkurchin thanks for editing this submission!  FYI this submission will be the second JOSS paper about pvlib.  The first was published in 2018 (https://github.com/openjournals/joss-reviews/issues/884).  This paper describes the substantial project growth over those ~5 years.

Looks like I need to tweak the submission file locations so that the build bot locates the correct set of files (right now it is using the files from the original 2018 paper).  One moment...
