### 2023-10-02 #5910 @rkurchin
👋 @brandons209, @EwaGomez, @shirubana, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
