### 2023-10-02 #5910 @rkurchin
@EwaGomez, great! Just to check though, you don't have any [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) with the authors?
