### 2023-10-02 #5910 @EwaGomez
@rkurchin No, there is no conflict of interest. The library is just closely related to my research topic, so I use it a lot. As you know, the author was recently a reviewer of my JOSS publication.
