+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -23,4 +23,4 @@ Documentation
 
 Contributors
 ~~~~~~~~~~~~
-
+* Arjan Keeman (:ghuser:`akeeman`)
