### 2023-10-04 2f136b6
Update pvlib/bifacial/infinite_sheds.py
Co-authored-by: Kevin Anderson <kevin.anderso@gmail.com>
