### 2023-10-04 1b6b511
Update infinite_sheds.py
