### 2023-10-04 f3a67da
Update infinite_sheds.py
fixed indentation issues
