### 2023-10-05 #5910 @dalonsoa
Hi @rkurchin , many thanks for the invite. Unfortunately, I cannot assume the review of this at the moment. If you need a suggestion for a possible reviewer, let me know and I'll point to to someone (also JOSS reviewer in the past).
