+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -29,3 +29,4 @@ Contributors
 * Arjan Keeman (:ghuser:`akeeman`)
 * Miguel Sánchez de León Peque (:ghuser:`Peque`)
 * Will Hobbs (:ghuser:`williamhobbs`)
+* Adam R. Jensen (:ghuser:`AdamRJensen`)
