### 2023-10-05 177e637
Update pvlib/irradiance.py
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
