+++ pvlib/temperature.py
@@ -535,7 +535,7 @@ def faiman_dyn(poa_global, temp_air, wind_speed=1.0, u0=25.0, u1=6.84,
     pvlib.temperature.faiman
 
     '''  # noQA: E501
-    
+
     if not isinstance(poa_global, pd.Series):
         raise ValueError('poa_global must be a pandas Series')
 
