### 2023-10-05 52c5580
Add Cliff's statement
