+++ pvlib/irradiance.py
@@ -542,8 +542,8 @@ def poa_components(aoi, dni, poa_sky_diffuse, poa_ground_diffuse):
 def get_ground_diffuse(surface_tilt, ghi, albedo=.25, surface_type=None):
     '''
     Estimate diffuse irradiance from ground reflections given
-    irradiance, albedo, and surface tilt. Isotropic ground reflections are
-    assumed.
+    irradiance, albedo, and surface tilt. The ground is assumed to
+    be isotropic and reflections are Lambertian.
 
     Function to determine the portion of irradiance on a tilted surface
     due to ground reflections. Any of the inputs may be DataFrames or
