+++ pvlib/temperature.py
@@ -563,8 +563,6 @@ def faiman_dyn(poa_global, temp_air, wind_speed=1.0, u0=25.0, u1=6.84,
     temp_difference = heat_input / total_loss_factor
     return temp_air + temp_difference
 
-    return temp_air
-
 
 def fit_faiman_dyn(temp_pv, poa_global, temp_air, wind_speed,
                    thermal_inertia=(0.0, 15.0, 1.0),
