### 2023-10-06 535da3f
create tools._logwrightomega function
