### 2023-10-06 09f140f
add method='logwright' in pvlib.pvsystem functions
