+++ docs/examples/operating-temperature/README.rst
@@ -1,0 +1,3 @@
+Simple dynamic version of the Faiman model
+------------------------------------------
+
