### 2023-10-09 ad57e0e
Rework example.
