### 2023-10-09 df66b39
Rework example some more.
