### 2023-10-09 b2f2269
More example tweaks.
