### 2023-10-10 7cc2ab8
Create requirements-py3.12.yml
