+++ ci/requirements-py3.12.yml
@@ -23,6 +23,6 @@ dependencies:
     - requests
     - scipy >= 1.4.0
     - statsmodels
-    - pip:
+    # - pip:
         # - nrel-pysam>=2.0  # not available for 3.12 as of 2023-10-10
         # - solarfactors  # required shapely<2 isn't available for 3.12
