### 2023-10-10 d004387
Remove model pseudo-option.
