### 2023-10-11 9dfc3d0
Update pvlib/iotools/solcast.py
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
