### 2023-10-11 85f8f73
Update pvlib/iotools/solcast.py
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
