### 2023-10-11 90983fe
Update pvlib/iotools/solcast.py
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
