### 2023-10-11 499f51a
midpoint docstring
