### 2023-10-16 fc1316c
improve martin_ruiz to physical, generalize tests
