### 2023-10-16 935443b
fix examples, split convert and fit examples
