### 2023-10-17 3475bf4
improve examples
