### 2023-10-17 f216d94
print parameters
