### 2023-10-17 ed35731
remove v0.10.2 whatsnew
