### 2023-10-17 fdcc952
Revert "remove v0.10.2 whatsnew"
This reverts commit ed357311b584392714dd8d86b36e64d9efb46349.
