+++ docs/sphinx/source/whatsnew/v0.10.2.rst
@@ -19,10 +19,6 @@ Enhancements
 * Added a continuous version of the Erbs diffuse-fraction/decomposition model.
   :py:func:`pvlib.irradiance.erbs_driesse` (:issue:`1755`, :pull:`1834`)
 * Added :py:func:`~pvlib.iam.interp` option as AOI losses model in
-  :py:class:`pvlib.modelchain.ModelChain` and
-  :py:class:`pvlib.pvsystem.PVSystem`. (:issue:`1742`, :pull:`1832`)
-* Added :py:func:`~pvlib.iam.convert` and :py:func:`~pvlib.iam.fit` that
-  convert between IAM models, and that fit an IAM model to data. (:issue:`1824`, :pull:`1827`)
   :py:class:`~pvlib.modelchain.ModelChain` and
   :py:class:`~pvlib.pvsystem.PVSystem`. (:issue:`1742`, :pull:`1832`)
 * :py:class:`~pvlib.pvsystem.PVSystem` objects with a single
@@ -80,8 +76,6 @@ Contributors
 * Anton Driesse (:ghuser:`adriesse`)
 * Lukas Grossar (:ghuser:`tongpu`)
 * Areeba Turabi (:ghuser:`aturabi`)
-* Cliff Hansen (:ghuser:`cwhanse`)
-
 * Saurabh Aneja (:ghuser:`spaneja`)
 * Miroslav Šedivý (:ghuser:`eumiro`)
 * kjsauer (:ghuser:`kjsauer`)
