### 2023-10-17 38bfb58
require scipy>=1.5.0
