+++ pvlib/tests/conftest.py
@@ -10,6 +10,7 @@ from functools import wraps
 
 import pvlib
 from pvlib.location import Location
+from pvlib.iam import _min_scipy
 
 pvlib_base_version = Version(Version(pvlib.__version__).base_version)
 
@@ -160,7 +161,6 @@ except ImportError:
 requires_pysam = pytest.mark.skipif(not has_pysam, reason="requires PySAM")
 
 
-from pvlib.iam import _min_scipy
 iam_scipy_ok = _min_scipy()
 
 requires_scipy_150 = pytest.mark.skipif(not iam_scipy_ok,
