+++ docs/sphinx/source/user_guide/roadmap.rst
@@ -1,0 +1,149 @@
+.. _roadmap:
+
+pvlib Roadmap
+=============
+
+**Last revision: October 2023**
+
+This document outlines a "wish list" for advancing pvlib-python, focusing
+on gaps in functionality and other code-related areas of improvement.
+
+Achieving these improvements will depend on community contributions.
+Therefore, it's not reasonable to set forth a timeline for this roadmap
+to be completed.
+For a more concrete plan of short-term additions to pvlib, please see
+the `milestone page <https://github.com/pvlib/pvlib-python/milestones>`_
+on GitHub.
+
+Additionally, an idea being listed here is no guarantee that it
+will ever be added to pvlib, nor is omission from this page an
+indication that an idea is not of interest.  This roadmap is an update
+of previous roadmaps based on requests from users:
+
+* `2023 pvlib-python user meeting notes
+  <https://github.com/pvlib/pvlib-python/wiki/2023-pvlib-python-user-meeting-notes>`_
+* `2022 pvlib development roadmap
+  <https://github.com/pvlib/pvlib-python/discussions/1581>`_
+* `2019 pvlib-python development roadmap
+  <https://github.com/pvlib/pvlib-python/wiki/2019-pvlib-python-Development-Roadmap>`_
+
+
+Core modeling capabilities
+--------------------------
+The core functionality of pvlib is its library of model functions,
+organized by the various conceptual steps of a PV system performance
+simulation.  pvlib has robust modeling capability in many of these topic
+areas, but coverage in some areas is poor or nonexistent.  Here is a
+partial list of modeling areas that could be improved:
+
+* Improved inverter performance models (power factor, off-MPP operation,
+  thermal derating, MPPT voltage range)
+* DC-DC optimizer performance models
+* Transformer models
+* Electrical mismatch functionality
+* Improved snow loss models, in particular for bifacial and tracked systems
+* Shade models
+* Degradation
+* More sophisticated single-axis tracker models
+* Bifacial irradiance models
+* Updated parameter libraries, e.g. the CEC PV Module database
+
+Note that published models may not exist for some of these enhancements,
+models would need to be developed and published before pvlib can
+provide a reference implementation.
+
+
+Auxiliary modeling capabilities
+-------------------------------
+Increasingly, pvlib is gaining functionality outside of the typical
+modeling workflow, for example parameter estimation/fitting, converting
+parameters between similar models, and clear-sky detection.
+These kinds of methods are not usually implemented in other
+performance modeling software, meaning providing pvlib implementations
+is of significant value.  However, making progress here is also somewhat
+constrained by the small number of existing published methods that can be
+implemented.
+
+
+System simulation
+-----------------
+pvlib's :py:class:`~pvlib.pvsystem.PVSystem` and
+:py:class:`~pvlib.modelchain.ModelChain` classes connect the individual
+model functions together to provide coherent system simulations.
+However, not all of pvlib's models are accessible with these classes.
+For example:
+
+* Bifacial irradiance models
+* Internal, external, and horizon shading
+* Soiling loss models
+
+In addition to more comprehensive integration of pvlib's model functions,
+the scope of these high-level capabilities could be expanded:
+
+* Multi-inverter systems
+* Geographic and electrical coordinate systems, to enable more detailed modeling
+* "out-of-code" model specification (e.g. configuration files)
+
+
+Data I/O
+--------
+:py:mod:`pvlib.iotools` provides functions to import
+ weather datasets, horizon profiles, component specifications,
+and other external data.  Expanding this capability by
+adding functions for additional data sources is of significant interest.
+:py:mod:`pvlib.iotools` would also benefit from further standardization
+(``map_variables``, etc) and more documentation.
+
[…]
