+++ docs/sphinx/source/user_guide/index.rst
@@ -17,4 +17,5 @@ User Guide
    comparison_pvlib_matlab
    variables_style_rules
    singlediode
+   roadmap
    faq
