+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -24,6 +24,7 @@ Testing
 
 Documentation
 ~~~~~~~~~~~~~
+* Create :ref:`roadmap` page. (:pull:`1893`)
 
 
 Contributors
