+++ pvlib/iotools/solcast.py
@@ -141,7 +141,8 @@ def get_solcast_historic(
     start : datetime-like
         First day of the requested period
     end : optional, datetime-like
-        Last day of the requested period. Must include one of ``end`` or ``duration``.
+        Last day of the requested period.
+        Must include one of ``end`` or ``duration``.
 
     duration : optional, default is None
         Must include either  ``end`` or ``duration``.
