### 2023-10-23 d8fd191
Update docs/sphinx/source/whatsnew.rst
Undoing changes to whatsnew.rst
