### 2023-10-23 63a2ca4
Fix plotting in plot_singlediode.py gallery page (#1895)
* Update plot_singlediode.py

fixed plot annotations by moving plt.show() further down

* Update whatsnew.rst

* Update v0.10.3.rst

* Update docs/sphinx/source/whatsnew.rst

Undoing changes to whatsnew.rst
