+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -24,7 +24,7 @@ Testing
 
 Documentation
 ~~~~~~~~~~~~~
-
+* Fixed a plotting issue in the IV curve gallery example (:pull:`1895`)
 
 Contributors
 ~~~~~~~~~~~~
