### 2023-10-27 #5994 @editorialbot
[REVIEW]: pvlib python: 2023 project update
**Submitting author:** @kandersolar (<a href="http://orcid.org/0000-0002-1166-7957">Kevin Anderson</a>)
**Repository:** https://github.com/pvlib/pvlib-python
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 0.10.3
