### 2023-10-28 469edc2
fix doi external links
Regex find: (?:doi|DOI):(?!`)\s?(.*?)(\.\n|\n)
Replace: :doi:`$1`$2
