### 2023-10-29 0ed672e
Cahnged expected reference in test_detect_clearskY_window to 1 from True to avoid Futurewarning
