### 2023-10-29 8ed03dd
Change reference to etr in ibird function to avoid FutureWarning
