+++ pvlib/tests/test_clearsky.py
@@ -855,7 +855,7 @@ def test_bird():
     # test scalars just at noon
     # XXX: calculations start at 12am so noon is at index = 12
     irrads3 = clearsky.bird(
-        zenith[12], airmass[12], aod_380nm, aod_500nm, h2o_cm, dni_extra=etr[12]
+        zenith[12], airmass[12], aod_380nm, aod_500nm, h2o_cm, dni_extra=etr.iloc[12]
     )
     Eb3, Ebh3, Gh3, Dh3 = (irrads3[_] for _ in field_names)
     # XXX: testdata starts at 1am so noon is at index = 11
