### 2023-10-29 12746bb
Update to iloc method for referencing by position in test_irradiance to get rid of FutureWarning
