### 2023-10-29 538f9c3
In test_singlediode change applymap to map to get rid of FutureWarning
