### 2023-10-31 57f67bb
Apply cwhanse suggestion
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
