### 2023-10-31 8794a2e
Apply cwhanse suggestion
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
