### 2023-10-31 6171e77
Modified cwhanse suggestion
Co-Authored-By: Cliff Hansen <5393711+cwhanse@users.noreply.github.com>
