### 2023-10-31 c75a662
Update modelchain.py
Co-Authored-By: Cliff Hansen <5393711+cwhanse@users.noreply.github.com>
