### 2023-10-31 9c52070
Some None's missed
F: (?<spaces> {8}| {4})(?<name>\w*) : [Nn]one[ ,]*(?<type>.*), optional$
R: $1$2 : $3, optional
Co-Authored-By: Cliff Hansen <5393711+cwhanse@users.noreply.github.com>
