### 2023-10-31 232156c
``none``'s, some more ``'s to variables and some other nones
Co-Authored-By: Cliff Hansen <5393711+cwhanse@users.noreply.github.com>
