+++ pvlib/iam.py
@@ -124,7 +124,7 @@ def physical(aoi, n=1.526, K=4.0, L=0.002, *, n_ar=None):
 
     n_ar : numeric, optional
         The effective index of refraction of the anti-reflective (AR) coating
-        (unitless). If n_ar is None (default), no AR coating is applied.
+        (unitless). If n_ar is not supplied, no AR coating is applied.
         A typical value for the effective index of an AR coating is 1.29.
 
     Returns
@@ -338,7 +338,7 @@ def martin_ruiz_diffuse(surface_tilt, a_r=0.16, c1=0.4244, c2=None):
     c2 : float
         Second fitting parameter for the expressions that approximate the
         integral of diffuse irradiance coming from different directions.
-        If c2 is None, it will be calculated according to the linear
+        If c2 is not specified, it will be calculated according to the linear
         relationship given in [3]_.
 
     Returns
