### 2023-10-31 b8f942b
Special ``none`` cases (please review)
Co-Authored-By: Cliff Hansen <5393711+cwhanse@users.noreply.github.com>
