### 2023-11-02 1a39119
Revert "Special ``none`` cases (please review)"
This reverts commit b8f942bb4396829986158d3995277c42f418e478.
