### 2023-11-02 eb0806d
Apply suggestions from code review
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
