### 2023-11-02 69c8321
Update pvlib/irradiance.py
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
