### 2023-11-06 1e412ce
Function prototype
