### 2023-11-06 a02bb78
Fix CAMS message error handler - Issue#1799
