### 2023-11-06 1da05c0
Fix CAMS message error handler - Issue#1799
