### 2023-11-06 e4a95a3
Update shading.rst
