+++ pvlib/shading.py
@@ -85,7 +85,7 @@ def masking_angle(surface_tilt, gcr, slant_height):
     ----------
     .. [1] D. Passias and B. Källbäck, "Shading effects in rows of solar cell
        panels", Solar Cells, Volume 11, Pages 281-291.  1984.
-       :doi:`10.1016/0379-6787(84)90017-6`
+       DOI: 10.1016/0379-6787(84)90017-6
     .. [2] Gilman, P. et al., (2018). "SAM Photovoltaic Model Technical
        Reference Update", NREL Technical Report NREL/TP-6A20-67399.
        Available at https://www.nrel.gov/docs/fy18osti/67399.pdf
@@ -167,7 +167,7 @@ def masking_angle_passias(surface_tilt, gcr):
     ----------
     .. [1] D. Passias and B. Källbäck, "Shading effects in rows of solar cell
        panels", Solar Cells, Volume 11, Pages 281-291.  1984.
-       :doi:`10.1016/0379-6787(84)90017-6`
+       DOI: 10.1016/0379-6787(84)90017-6
     """
     # wrap it in an array so that division by zero is handled well
     beta = np.radians(np.array(surface_tilt))
@@ -226,7 +226,7 @@ def sky_diffuse_passias(masking_angle):
     ----------
     .. [1] D. Passias and B. Källbäck, "Shading effects in rows of solar cell
        panels", Solar Cells, Volume 11, Pages 281-291.  1984.
-       :doi:`10.1016/0379-6787(84)90017-6`
+       DOI: 10.1016/0379-6787(84)90017-6
     .. [2] Gilman, P. et al., (2018). "SAM Photovoltaic Model Technical
        Reference Update", NREL Technical Report NREL/TP-6A20-67399.
        Available at https://www.nrel.gov/docs/fy18osti/67399.pdf
@@ -234,111 +234,25 @@ def sky_diffuse_passias(masking_angle):
     return 1 - cosd(masking_angle/2)**2
 
 
-def projected_solar_zenith_angle(solar_zenith, solar_azimuth,
-                                 axis_tilt, axis_azimuth):
+def projected_solar_zenith_angle(apparent_zenith, azimuth):
     r"""
     Calculate projected solar zenith angle in degrees.
 
-    This solar zenith angle is projected onto the plane whose normal vector is
-    defined by ``axis_tilt`` and ``axis_azimuth``. The normal vector is in the
-    direction of ``axis_azimuth`` (clockwise from north) and tilted from
-    horizontal by ``axis_tilt``. See Figure 5 in [1]_:
-
-    .. figure:: ../../_images/Anderson_Mikofski_2020_Fig5.jpg
-       :alt: Wire diagram of coordinates systems to obtain the projected angle.
-       :align: center
-       :scale: 50 %
-
-       Fig. 5, [1]_: Solar coordinates projection onto tracker rotation plane.
-
     Parameters
     ----------
-    solar_zenith : numeric
+    apparent_zenith : numeric
         Sun's apparent zenith in degrees.
-    solar_azimuth : numeric
+    azimuth : numeric
         Sun's azimuth in degrees.
-    axis_tilt : numeric
-        Axis tilt angle in degrees. From horizontal plane to array plane.
-    axis_azimuth : numeric
-        Axis azimuth angle in degrees.
-        North = 0°; East = 90°; South = 180°; West = 270°
 
     Returns
     -------
     Projected_solar_zenith : numeric
         In degrees.
-
-    Notes
-    -----
-    This projection has a variety of applications in PV. For example:
-
-    - Projecting the sun's position onto the plane perpendicular to
-      the axis of a single-axis tracker (i.e. the plane
-      whose normal vector coincides with the tracker torque tube)
[…]
