+++ pvlib/tests/test_shading.py
@@ -4,29 +4,27 @@ import pandas as pd
 from pandas.testing import assert_series_equal
 from numpy.testing import assert_allclose
 import pytest
-from datetime import timezone, timedelta
 
 from pvlib import shading
 
 
 @pytest.fixture
 def test_system():
-    syst = {
-        "height": 1.0,
-        "pitch": 2.0,
-        "surface_tilt": 30.0,
-        "surface_azimuth": 180.0,
-        "rotation": -30.0,
-    }  # rotation of right edge relative to horizontal
-    syst["gcr"] = 1.0 / syst["pitch"]
+    syst = {'height': 1.0,
+            'pitch': 2.,
+            'surface_tilt': 30.,
+            'surface_azimuth': 180.,
+            'rotation': -30.}  # rotation of right edge relative to horizontal
+    syst['gcr'] = 1.0 / syst['pitch']
     return syst
 
 
 def test__ground_angle(test_system):
     ts = test_system
-    x = np.array([0.0, 0.5, 1.0])
-    angles = shading.ground_angle(ts["surface_tilt"], ts["gcr"], x)
-    expected_angles = np.array([0.0, 5.866738789543952, 9.896090638982903])
+    x = np.array([0., 0.5, 1.0])
+    angles = shading.ground_angle(
+        ts['surface_tilt'], ts['gcr'], x)
+    expected_angles = np.array([0., 5.866738789543952, 9.896090638982903])
     assert np.allclose(angles, expected_angles)
 
 
@@ -40,7 +38,7 @@ def test__ground_angle_zero_gcr():
 
 @pytest.fixture
 def surface_tilt():
-    idx = pd.date_range("2019-01-01", freq="h", periods=3)
+    idx = pd.date_range('2019-01-01', freq='h', periods=3)
     return pd.Series([0, 20, 90], index=idx)
 
 
@@ -109,117 +107,13 @@ def test_sky_diffuse_passias_scalar(average_masking_angle, shading_loss):
         assert np.isclose(loss, actual_loss)
 
 
-@pytest.fixture
-def true_tracking_angle_and_inputs_NREL():
-    # data from NREL 'Slope-Aware Backtracking for Single-Axis Trackers'
-    # doi.org/10.2172/1660126 ; Accessed on 2023-11-06.
-    tzinfo = timezone(timedelta(hours=-5))
-    axis_tilt_angle = 9.666  # deg
-    axis_azimuth_angle = 195.0  # deg
-    timedata = pd.DataFrame(
-        columns=("Apparent Elevation", "Solar Azimuth", "True-Tracking"),
-        data=(
-            (2.404287, 122.791770, -84.440),
-            (11.263058, 133.288729, -72.604),
-            (18.733558, 145.285552, -59.861),
-            (24.109076, 158.939435, -45.578),
-            (26.810735, 173.931802, -28.764),
-            (26.482495, 189.371536, -8.475),
-            (23.170447, 204.136810, 15.120),
-            (17.296785, 217.446538, 39.562),
-            (9.461862, 229.102218, 61.587),
-            (0.524817, 239.330401, 79.530),
-        ),
-    )
-    timedata.index = pd.date_range(
-        "2019-01-01T08", "2019-01-01T17", freq="1H", tz=tzinfo
-    )
-    timedata["Apparent Zenith"] = 90.0 - timedata["Apparent Elevation"]
-    return (axis_tilt_angle, axis_azimuth_angle, timedata)
-
-
-@pytest.fixture
-def projected_solar_zenith_angle_edge_cases():
-    premises_and_result_matrix = pd.DataFrame(
-        data=[
-            # s_zen | s_azm | ax_tilt | ax_azm | psza
-            [   0,       0,      0,        0,      0],
[…]
