### 2023-11-06 64c3217
Substitute changing to float64 dtype using map with base functionality that's accessible across Pandas versions
