+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -32,3 +32,4 @@ Contributors
 * Miguel Sánchez de León Peque (:ghuser:`Peque`)
 * Will Hobbs (:ghuser:`williamhobbs`)
 * Anton Driesse (:ghuser:`adriesse`)
+* :ghuser:`matsuobasho`
