### 2023-11-06 8cfcedc
Update line break in test_clearsky to adhere to line length limit
