+++ pvlib/iotools/sodapro.py
@@ -217,11 +217,12 @@ def get_cams(latitude, longitude, start, end, email, identifier='mcclear',
 
     # Response from CAMS follows the status and reason format of PyWPS4
     # If an error occurs on our side, we will return error 400 - bad request
-    # Additional information is available in the response text so we add it here 
+    # Additional information is available in the response text so
+    # we add it here
     # to the error displayed to facilitate users effort to fix their request
     if not res.ok:
         errors = res.text.split('ows:ExceptionText')[1][1:-2]
-        res.reason = "%s: <%s>"%(res.reason, errors)
+        res.reason = "%s: <%s>" % (res.reason, errors)
         res.raise_for_status()
     # Successful requests returns a csv data file
     elif res.headers['Content-Type'] == 'application/csv':
