### 2023-11-08 2a72f33
Implementation
From NREL paper
