### 2023-11-08 f012aae
Extend tests (compare with singleaxis) & format with ruff
