+++ pvlib/shading.py
@@ -277,8 +277,10 @@ def projected_solar_zenith_angle(surface_tilt, surface_azimuth,
     sy = cosd(solar_apparent_elevation) * cosd(solar_azimuth)
     sz = sind(solar_apparent_elevation)
     sx_prime = sx * cosd(surface_azimuth) - sy * sind(surface_azimuth)
-    sz_prime = (sx * sind(surface_azimuth) * sind(surface_tilt)
-                + sy * sind(surface_tilt) * cosd(surface_azimuth)
-                + sz * cosd(surface_tilt))
+    sz_prime = (
+        sx * sind(surface_azimuth) * sind(surface_tilt)
+        + sy * sind(surface_tilt) * cosd(surface_azimuth)
+        + sz * cosd(surface_tilt)
+    )
     theta_T = np.degrees(np.arctan2(sx_prime, sz_prime))
     return theta_T
