### 2023-11-09 aee7b7d
Apply suggestions from code review
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
