### 2023-11-09 e6a49cd
Fix associated UT and add contribution in WhatsNews
