+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -23,6 +23,8 @@ Enhancements
 * Add :py:func:`pvlib.iotools.read_solaranywhere` and
   :py:func:`pvlib.iotools.get_solaranywhere` for reading and retrieving
   SolarAnywhere solar irradiance data. (:pull:`1497`, :discuss:`1310`)
+* Added function :py:func:`pvlib.shading.projected_solar_zenith_angle`,
+  a common calculation in shading and tracking. (:issue:`1734`, :pull:`1904`)
 
 Bug fixes
 ~~~~~~~~~
@@ -73,3 +75,4 @@ Contributors
 * Mark Mikofski (:ghuser:`mikofski`)
 * Phoebe Pearce (:ghuser:`phoebe-p`)
 * Eva-Maria Grommes (:ghuser:`EwaGomez`)
+* Echedey Luis (:ghuser:`echedey-ls`)
