### 2023-11-09 607a246
Minor Python optimization a la tracking.singleaxis
