### 2023-11-09 79b5f4f
Typo found by Mikofski
Reported at: https://github.com/pvlib/pvlib-python/pull/1725#discussion_r1190610810

Confirmed via "Slope-Aware Backtracking for Single-Axis Trackers", paragraph after Eq. 1

Co-Authored-By: Mark Mikofski <bwana.marko@yahoo.com>
