### 2023-11-09 88cbfc0
Elevation -> Zenith
Co-Authored-By: Kevin Anderson <57452607+kandersolar@users.noreply.github.com>
