+++ docs/sphinx/source/user_guide/weather_data.rst
@@ -7,46 +7,70 @@ Simulating the performance of a PV system requires irradiance and meteorological
 as the inputs to a PV system model.  Weather datasets are available
 from many sources and in many formats.  The :py:mod:`pvlib.iotools` module
 contains functions to easily retrieve and import such datasets in a standardized
-form that is convenient to use with the rest of pvlib.  For a complete list
-of functions related to retrieving and importing weather data, see :ref:`iotools`.
+form that is convenient to use with the rest of pvlib.  
+
+The primary focus of :py:mod:`pvlib.iotools` is time series solar resource
+data like the datasets from PVGIS and the NSRDB, but it also provides
+functionality for other types of data useful for certain aspects of PV modeling
+(e.g. precipitation data from :py:func:`~pvlib.iotools.get_acis_prism`
+for soiling modeling, and horizon profiles from :py:func:`~pvlib.iotools.get_pvgis_horizon`
+for horizon shade modeling).
+
+For a complete list of functions related to retrieving and importing weather
+data, see :ref:`iotools`.
 
 
 Types of weather data sources
 -----------------------------
 
-Ground station measurements
-***************************
+Weather data can be grouped into a few fundamental categories.  Which
+type is most useful depends on the application.  Here we provide a high-level
+overview of different types of weather data, and when you might want to use
+them.
 
-From in-situ monitoring equipment. If properly maintained and quality-controlled,
-these are the highest quality source of weather information. However, the coverage
-depends on a weather station having been set up in advance for the location and
-times of interest. There are datasets such as BSRN and SURFRAD which make their
-measurement data publicly available.
+1. **Ground station measurements**:
+   From in-situ monitoring equipment. If properly maintained and
+   quality-controlled, these are the highest quality
+   source of weather information. However, the coverage depends on
+   a weather station having been set up in advance for the location and
+   times of interest. Some ground station networks like the BSRN and SURFRAD
+   make their measurement data publicly available.
+   
+   Data from public ground station measurement networks are useful if you
+   want accurate, high-resolution data but have flexibility around the
+   specific measurement location.
 
+2. **Satellite data**: 
+   These sources process satellite imagery (typically from geostationary
+   satellites) to identify and classify clouds, and combine this with solar
+   irradiance models to produce irradiance estimates. The quality is
+   generally much higher than NWP, but still not as good as a well-maintained
+   weather station. They have high spatial and temporal resolution
+   corresponding to the source satellite imagery, and are generally
+   optimised to estimate solar irradiance for PV applications. Free sources
+   such as PVGIS are available, and commerical sources such as SolarAnywhere,
+   Solcast and Solargis provide paid options though often have free trials.
+   
+   Satellite data is useful when suitable ground station measurements are
+   not available for the location and/or times of interest.
 
-Numerical Weather Prediction (NWP)
-**********************************
+3. **Numerical Weather Prediction (NWP)**:
+   These are mathematical simulations of weather systems.
+   The data quality is much lower than that of measurements and
+   satellite data, owing in part to coarser spatial and temporal
+   resolution, as well as many models not being optimised for solar
+   irradiance for PV applications. On the plus side, these models typically
+   have worldwide coverage, with some regional models (e.g. HRRR) sacrifice
[…]
