+++ paper/paper.bib
@@ -80,11 +80,17 @@
   institution = {The Linux Foundation}
 }
 
-@article{pvlibiotools,
+@article{Jensen2023,
+  doi = {10.1016/j.solener.2023.112092},
+  url = {https://doi.org/10.1016/j.solener.2023.112092},
+  year = {2023},
+  month = dec,
+  publisher = {Elsevier {BV}},
+  volume = {266},
+  pages = {112092},
   author = {Adam R. Jensen and Kevin S. Anderson and William F. Holmgren and Mark A. Mikofski and Clifford W. Hansen and Leland J. Boeman and Roel Loonen},
-  title = {{pvlib} iotools - open-source Python functions for seamless access to solar irradiance data},
-  year = {submitted},
-  journal = {Solar Energy},
+  title = {{pvlib} iotools{\textemdash}Open-source Python functions for seamless access to solar irradiance data},
+  journal = {Solar Energy}
 }
 
 @Article{         numpy,
