### 2023-11-12 662846e
Add reverse transposition function and two helpers.
