### 2023-11-12 2c4e4b9
Minor doc string fixes.
