### 2023-11-13 pvlib/pvlib-python#1908 @EwaGomez
JOSS-review: statement of need
While working through the [JOSS publication](https://github.com/openjournals/joss-reviews/issues/5994) checklist, I was looking for a statement of need for pvlib in the documentation. You mention the statement of need in the package overview introduction, but I think it would be interesting to clarify the target audience on the main landing page of the documentation as well.
