### 2023-11-13 pvlib/pvlib-python#1909 @EwaGomez
JOSS-review: other PV modeling tools
In your [JOSS ](https://github.com/openjournals/joss-reviews/issues/5994) paper you state that compared to other PV modelling tools, pvlib stands out in several key aspects. Could you name and cite some of the tools and link them to the aspects where pvlib stands out?
