### 2023-11-13 9bf9064
Add full_output option similar to newton().
