### 2023-11-13 97d4559
Some improvements to the two examples.
