### 2023-11-14 e5025b4
Refine examples and fix test.
