### 2023-11-14 58e8ded
Try to get rid of matplotlib warning in example.
