+++ docs/examples/irradiance-transposition/plot_rtranpose_year.py
@@ -26,7 +26,6 @@ Author: Anton Driesse
 import os
 import pandas as pd
 
-import matplotlib
 import matplotlib.pyplot as plt
 
 import pvlib
