+++ paper/paper.bib
@@ -66,10 +66,81 @@
 
 @inproceedings{Holmgren2018,
 	Author = {William F. Holmgren and Clifford W. Hansen and Joshua S. Stein and Mark A. Mikofski},
-	Booktitle = {IEEE 45th Photovoltaic Specialists Conference},
+	Booktitle = {45th Photovoltaic Specialists Conference},
 	Title = {Review of open source tools for PV modeling},
 	Year = {2018},
 	doi = {10.5281/zenodo.1401378}
+}
+
+@inproceedings{Andrews2014,
+  doi = {10.1109/pvsc.2014.6925501},
+  url = {https://doi.org/10.1109/pvsc.2014.6925501},
+  year = {2014},
+  month = jun,
+  publisher = {{IEEE}},
+  author = {Robert W. Andrews and Joshua S. Stein and Clifford Hansen and Daniel Riley},
+  title = {Introduction to the open source {PV} {LIB} for python Photovoltaic system modelling package},
+  booktitle = {40th Photovoltaic Specialist Conference}
+}
+
+@inproceedings{Mermoud1994,
+  title = {{PVSYST}: a user-friendly software for {PV}-systems simulation},
+  booktitle = {Twelfth European Photovoltaic Solar Energy Conference},
+  year = {1994},
+  author = {André Mermoud}
+}
+
+@inproceedings{Mikofski2018,
+  title = {Accurate Performance Predictions of Large PV Systems with Shading using Submodule Mismatch Calculation},
+  url = {http://dx.doi.org/10.1109/PVSC.2018.8547323},
+  DOI = {10.1109/pvsc.2018.8547323},
+  booktitle = {7th World Conference on Photovoltaic Energy Conversion (WCPEC)},
+  publisher = {IEEE},
+  author = {Mikofski,  Mark A. and Lynn,  Matthew and Byrne,  James and Hamer,  Mike and Neubert,  Anja and Newmiller,  Jeff},
+  year = {2018},
+  month = jun 
+}
+
+@inproceedings{Passow2017,
+  title = {PlantPredict: Solar Performance Modeling Made Simple},
+  url = {http://dx.doi.org/10.1109/PVSC.2017.8366450},
+  DOI = {10.1109/pvsc.2017.8366450},
+  booktitle = {44th Photovoltaic Specialist Conference (PVSC)},
+  publisher = {IEEE},
+  author = {Passow,  Kendra and Ngan,  Lauren and Rich,  Geoffrey and Lee,  Mitch and Kaplan,  Stephen},
+  year = {2017},
+  month = jun 
+}
+
+@techreport{Gilman2018,
+  title = {SAM Photovoltaic Model Technical Reference 2016 Update},
+  url = {http://dx.doi.org/10.2172/1429291},
+  DOI = {10.2172/1429291},
+  institution = {Office of Scientific and Technical Information (OSTI)},
+  author = {Gilman,  Paul and DiOrio,  Nicholas A. and Freeman,  Janine M. and Janzou,  Steven and Dobos,  Aron and Ryberg,  David},
+  year = {2018},
+  month = mar 
+}
+
+@misc{pysam,
+  doi = {10.11578/DC.20190903.1},
+  url = {https://www.osti.gov/doecode/biblio/29023},
+  author = {Gilman,  Paul and Janzou,  Steven and Guittet,  Darice and Freeman,  Janine and DiOrio,  Nicholas and Blair,  Nathan and Boyd,  Matthew and Neises,  Ty and Wagner,  Michael},
+  language = {en},
+  title = {PySAM (Python Wrapper for System Advisor Model "SAM") [SWR-19-57]},
[…]
