### 2023-11-15 320886c
edit project description in README, RTD, PyPI
