+++ docs/sphinx/source/index.rst
@@ -6,12 +6,14 @@ pvlib python
 
 pvlib python is a community supported tool that provides a set of
 functions and classes for simulating the performance of photovoltaic
-energy systems. pvlib python was originally ported from the PVLIB MATLAB
-toolbox developed at Sandia National Laboratories and it implements many
-of the models and methods developed at the Labs. More information on
-Sandia Labs PV performance modeling programs can be found at the `PVPMC <https://pvpmc.sandia.gov/>`_. We collaborate with the PVLIB MATLAB project,
-but operate independently of it.
+energy systems.  The core mission of pvlib python is to provide open,
+reliable, interoperable, and benchmark implementations of PV system models.
 
+pvlib python started out as a Python translation of the PVLIB MATLAB
+toolbox (developed by the `PVPMC <https://pvpmc.sandia.gov/>`_ at
+Sandia National Laboratories) in 2013 and has grown substantially since then.
+Today it contains code contributions from over a hundred individuals worldwide
+and is maintained by a core group of PV modelers from a variety institutions.
 The source code for pvlib python is hosted on `github
 <https://github.com/pvlib/pvlib-python>`_.
 
