### 2023-11-15 pvlib/pvlib-python#1908 @kandersolar
That landing page text is long overdue for some edits :)

#1913 goes some way towards addressing this issue.  Please let me know if you think more clarification would be helpful!
