### 2023-11-16 pvlib/pvlib-python#1909 @EwaGomez
Thank you for your edits!
