### 2023-11-17 44cd666
Apply suggestions from code review
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
