+++ paper/paper.bib
@@ -66,7 +66,7 @@
 
 @inproceedings{Holmgren2018,
 	Author = {William F. Holmgren and Clifford W. Hansen and Joshua S. Stein and Mark A. Mikofski},
-	Booktitle = {45th Photovoltaic Specialists Conference},
+	Booktitle = {45th IEEE Photovoltaic Specialists Conference},
 	Title = {Review of open source tools for PV modeling},
 	Year = {2018},
 	doi = {10.5281/zenodo.1401378}
