### 2023-11-17 9d29042
revision re: other open-source projects
