### 2023-11-17 7d2673e
bibtex tweaks
