+++ paper/paper.bib
@@ -80,7 +80,7 @@
   publisher = {{IEEE}},
   author = {Robert W. Andrews and Joshua S. Stein and Clifford Hansen and Daniel Riley},
   title = {Introduction to the open source {PV} {LIB} for python Photovoltaic system modelling package},
-  booktitle = {40th Photovoltaic Specialist Conference}
+  booktitle = {40th IEEE Photovoltaic Specialists Conference}
 }
 
 @inproceedings{Mermoud1994,
@@ -105,7 +105,7 @@
   title = {PlantPredict: Solar Performance Modeling Made Simple},
   url = {http://dx.doi.org/10.1109/PVSC.2017.8366450},
   DOI = {10.1109/pvsc.2017.8366450},
-  booktitle = {44th Photovoltaic Specialist Conference (PVSC)},
+  booktitle = {44th IEEE Photovoltaic Specialists Conference (PVSC)},
   publisher = {IEEE},
   author = {Passow,  Kendra and Ngan,  Lauren and Rich,  Geoffrey and Lee,  Mitch and Kaplan,  Stephen},
   year = {2017},
@@ -136,7 +136,7 @@
   title = {Introducing CASSYS: An open-source software for simulation of grid-connected photovoltaic systems},
   url = {http://dx.doi.org/10.1109/PVSC.2016.7749839},
   DOI = {10.1109/pvsc.2016.7749839},
-  booktitle = {43rd Photovoltaic Specialists Conference},
+  booktitle = {43rd IEEE Photovoltaic Specialists Conference},
   publisher = {IEEE},
   author = {Pai,  Abhijeet and Thevenard,  Didier},
   year = {2016},
