### 2023-11-21 #5994 @phoebe-p
Hi @kandersolar, this is not a major issue but as a Mac user I (briefly, before I remembered the quirks of zsh) ran into the issue from this (closed) [PR](https://github.com/pvlib/pvlib-python/pull/1167). Since zsh is the default shell on all MacOS now as far as I know, there may be quite a few people running into this - it's pretty easy to Google, but perhaps it could be added to the FAQ or similar?
