### 2023-11-21 232f4f5
Add window_length keyword argument to detect_clearsky example in clearsky.rst user guide
