### 2023-11-21 091d47b
Apply suggestions from code review
Co-authored-by: Anton Driesse <anton.driesse@pvperformancelabs.com>
