+++ docs/sphinx/source/index.rst
@@ -4,7 +4,7 @@
 pvlib python
 ============
 
-pvlib python is a community supported tool that provides a set of
+pvlib python is a community supported toolbox that provides a set of
 functions and classes for simulating the performance of photovoltaic
 energy systems.  The core mission of pvlib python is to provide open,
 reliable, interoperable, and benchmark implementations of PV system models.
