+++ README.md
@@ -69,9 +69,9 @@
 </table>
 
 
-pvlib python is a community supported tool that provides a set of
+pvlib python is a community supported toolbox that provides a set of
 functions and classes for simulating the performance of photovoltaic
-energy systems.  The core mission of pvlib python is to provide open,
+energy systems and accomplishing related tasks.  The core mission of pvlib python is to provide open,
 reliable, interoperable, and benchmark implementations of PV system models.
 
 pvlib python started out as a Python translation of the PVLIB MATLAB
