### 2023-11-21 dfebe45
adopt reviewer suggestions
Co-Authored-By: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
Co-Authored-By: Will Holmgren <william.holmgren@gmail.com>
