### 2023-11-21 fe6ba59
Removed cs.index and window_length arguments from clearsky.detect_clearsky in clearsky.rst user guide
