+++ docs/sphinx/source/user_guide/clearsky.rst
@@ -642,8 +642,7 @@ Now we run the synthetic data and clear sky estimate through the
 
 .. ipython:: python
 
-    clear_samples = clearsky.detect_clearsky(ghi, cs['ghi'], cs.index,
-                     window_length=10)
+    clear_samples = clearsky.detect_clearsky(ghi, cs['ghi'])
 
     fig, ax = plt.subplots()
 
