### 2023-11-21 892dd0e
Update geographical coverage description
