### 2023-11-21 37adc7d
Apply suggestions from code review
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
