### 2023-11-21 dd68286
Fix error in clear-sky user's guide docs page (#1915)
* Add window_length keyword argument to detect_clearsky example in clearsky.rst user guide

* Update v0.10.03.rst "what's new" file to reflect change

* Removed cs.index and window_length arguments from clearsky.detect_clearsky in clearsky.rst user guide
