+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -25,6 +25,7 @@ Testing
 Documentation
 ~~~~~~~~~~~~~
 * Fixed a plotting issue in the IV curve gallery example (:pull:`1895`)
+* Fixed `detect_clearsky` example in `clearsky.rst` (:issue:`1914`)
 
 Contributors
 ~~~~~~~~~~~~
