+++ pvlib/iotools/sodapro.py
@@ -59,7 +59,7 @@ def get_cams(latitude, longitude, start, end, email, identifier='mcclear',
     Requests: max. 100 per day
 
     Geographical coverage: worldwide for CAMS McClear and approximately -66° to
-    66° in latitude and -66° to 180° in longitude for CAMS Radiation. See [3]_
+    66° in latitude and -66° to 200° in longitude for CAMS Radiation. See [3]_
     for a map of the geographical coverage.
 
     Parameters
