### 2023-11-22 5394cb1
Create weather data/iotools page in User's Guide (#1754)
* add a bit of content

* Apply suggestions from code review

Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>

* revisions

* further improvements

* Apply suggestions from code review

Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>

* make :py:mod:`pvlib.iotools` link to API reference

* move mention of iotools paper

* cleanup

* whatsnew

---------

Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
