+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -24,6 +24,7 @@ Testing
 
 Documentation
 ~~~~~~~~~~~~~
+* Create :ref:`weatherdata` User's Guide page. (:pull:`1754`)
 * Fixed a plotting issue in the IV curve gallery example (:pull:`1895`)
 * Fixed `detect_clearsky` example in `clearsky.rst` (:issue:`1914`)
 
@@ -34,3 +35,6 @@ Contributors
 * Will Hobbs (:ghuser:`williamhobbs`)
 * Anton Driesse (:ghuser:`adriesse`)
 * :ghuser:`matsuobasho`
+* Harry Jack (:ghuser:`harry-solcast`)
+* Adam R. Jensen (:ghuser:`AdamRJensen`)
+* Kevin Anderson (:ghuser:`kandersolar`)
