### 2023-11-24 pvlib/pvlib-python#1908 @EwaGomez
Yes, it does! I will finish my review now :-) Thanks!
