### 2023-11-24 #5994 @EwaGomez
From my side the review is complete. The author solved both issues, I opened.
