### 2023-11-27 6e7ec7b
Update pvlib/irradiance.py
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
