+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -29,6 +29,7 @@ Documentation
 ~~~~~~~~~~~~~
 * Fixed a plotting issue in the IV curve gallery example (:pull:`1895`)
 * Added two examples to demonstrate reverse transposition (:pull:`1907`)
+* Fixed `detect_clearsky` example in `clearsky.rst` (:issue:`1914`)
 
 Contributors
 ~~~~~~~~~~~~
