+++ pvlib/tests/test_iam.py
@@ -8,7 +8,7 @@ import numpy as np
 import pandas as pd
 
 import pytest
-from .conftest import (assert_series_equal, requires_scipy_150)
+from .conftest import assert_series_equal, requires_scipy_150
 from numpy.testing import assert_allclose
 
 from pvlib import iam as _iam
