### 2023-11-28 2a0b815
edit both notes
