+++ pvlib/iam.py
@@ -1174,9 +1174,8 @@ def convert(source_name, source_params, target_name, weight=None, fix_n=True,
 
     .. math::
 
-        \Sum_{\theta=0}^90 weight \times \\left(\theta \\right)
-        * \\left(\\| source \\left(\theta \\right)
-        - target \\left(\theta \\right) \\| \\right)
+        \\Sum_{\\theta=0}^{90} weight \\left(\\theta \\right) \\times 
+        \\| source \\left(\\theta \\right) - target \\left(\\theta \\right) \\|
         
     References
     ----------
@@ -1278,6 +1277,15 @@ def fit(measured_aoi, measured_iam, model_name, weight=None, xtol=None):
     .. [1] Jones, A. R., Hansen, C. W., Anderson, K. S. Parameter estimation
        for incidence angle modifier models for photovoltaic modules. Sandia
        report SAND2023-13944 (2023).
+
+    Note
+    ----
+    Model parameters are determined by minimizing
+
+    .. math::
+
+        \\Sum_{measured AOI} weight \\left( AOI \\right) \\times 
+        \\| measured IAM \\left( AOI \\right) - model \\left(\\AOI \\right) \\|
 
     See Also
     --------
