### 2023-11-28 4e165a0
polish the notes
