### 2023-11-29 6f135ee
remove history and support statement from setup.py
