### 2023-11-29 ba67502
README: move history/funding to separate section
