+++ README.md
@@ -74,16 +74,6 @@ functions and classes for simulating the performance of photovoltaic
 energy systems and accomplishing related tasks.  The core mission of pvlib python is to provide open,
 reliable, interoperable, and benchmark implementations of PV system models.
 
-pvlib python started out as a Python translation of the PVLIB MATLAB
-toolbox (developed by the [PVPMC](https://pvpmc.sandia.gov/) at
-Sandia National Laboratories) in 2013 and has grown substantially since then.
-Today it contains code contributions from over a hundred individuals worldwide
-and is maintained by a core group of PV modelers from a variety institutions.
-
-pvlib has been supported directly and indirectly by DOE, NumFOCUS, and
-Google Summer of Code funding, companies that allow their employees to
-contribute, and from personal time.
-
 
 Documentation
 =============
@@ -146,6 +136,20 @@ License
 =======
 
 BSD 3-clause.
+
+
+History and acknowledgement
+===========================
+
+pvlib python started out as a Python translation of the PVLIB MATLAB
+toolbox (developed by the `PVPMC <https://pvpmc.sandia.gov/>`_ at
+Sandia National Laboratories) in 2013 and has grown substantially since then.
+Today it contains code contributions from over a hundred individuals worldwide
+and is maintained by a core group of PV modelers from a variety institutions.
+
+pvlib has been supported directly and indirectly by DOE, NumFOCUS, and
+Google Summer of Code funding, university research projects,
+companies that allow their employees to contribute, and from personal time.
 
 
 NumFOCUS
