+++ paper/paper.bib
@@ -18,9 +18,10 @@
 @inproceedings{pvpmc2022_pvlib_update,
     title = {{pvlib} python 2022 update},
     author = {Anderson, Kevin and Holmgren, Will and Hansen, Clifford and Mikofski, Mark and Jensen, Adam R. and Driesse, Anton},
-    url = {https://www.osti.gov/biblio/1886878},
+    url = {https://doi.org/10.2172/2004407},
     booktitle = {PV Performance Modeling and Monitoring Workshop},
     year = {2022},
+    doi = {10.2172/2004407},
 }
 
 @inproceedings{pvpmc2022_pvanalytics_update,
