### 2023-11-29 c78ec3b
Edit project description in README, RTD, PyPI (#1913)
* edit project description in README, RTD, PyPI

* Apply suggestions from code review

Co-authored-by: Anton Driesse <anton.driesse@pvperformancelabs.com>

* adopt reviewer suggestions

Co-Authored-By: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
Co-Authored-By: Will Holmgren <william.holmgren@gmail.com>

* remove history and support statement from setup.py

* add "university research projects"

* README: move history/funding to separate section

---------

Co-authored-by: Anton Driesse <anton.driesse@pvperformancelabs.com
[…]
