+++ setup.py
@@ -12,14 +12,11 @@ except ImportError:
 DESCRIPTION = ('A set of functions and classes for simulating the ' +
                'performance of photovoltaic energy systems.')
 LONG_DESCRIPTION = """
-PVLIB Python is a community supported tool that provides a set of
+pvlib python is a community developed toolbox that provides a set of
 functions and classes for simulating the performance of photovoltaic
-energy systems. PVLIB Python was originally ported from the PVLIB MATLAB
-toolbox developed at Sandia National Laboratories and it implements many
-of the models and methods developed at the Labs. More information on
-Sandia Labs PV performance modeling programs can be found at
-https://pvpmc.sandia.gov/. We collaborate with the PVLIB MATLAB project,
-but operate independently of it.
+energy systems and accomplishing related tasks.  The core mission of pvlib
+python is to provide open, reliable, interoperable, and benchmark
+implementations of PV system models.
 
 We need your help to make pvlib-python a great tool!
 
