### 2023-11-29 #5994 @kandersolar
Thank you @EwaGomez and @phoebe-p for the constructive feedback!

@rkurchin I think we are ready to proceed here.  Note that the suggested DOI for "pytest" in the missing DOI list above is not related to what we are trying to cite, so I have left that one unaddressed.
