### 2023-11-29 34ee1fe
increase minimum scipy to 1.5.0
