+++ ci/requirements-py3.8.yml
@@ -21,7 +21,7 @@ dependencies:
     - python=3.8
     - pytz
     - requests
-    - scipy >= 1.4.0
+    - scipy >= 1.5.0
     - statsmodels
     - pip:
         - nrel-pysam>=2.0
