### 2023-11-29 dba9a9c
increase minimum scipy to 1.5.0 (#1919)
