### 2023-11-29 32aa64a
remove test for scipy
