+++ pvlib/iam.py
@@ -1094,18 +1094,6 @@ def _process_return(target_name, optimize_result):
     return target_params
 
 
-def _min_scipy():
-    '''convert and fit require scipy>=1.5.0
-    '''
-    from scipy import __version__
-    major, minor = __version__.split('.')[:2]
-    if (int(major) >= 1) & (int(minor) >= 5):
-        return True
-    else:
-        warnings.warn('iam.convert and iam.fit require scipy>=1.5.0')
-        return False
-
-
 def convert(source_name, source_params, target_name, weight=_sin_weight,
             fix_n=True, xtol=None):
     """
@@ -1192,43 +1180,38 @@ def convert(source_name, source_params, target_name, weight=_sin_weight,
     pvlib.iam.martin_ruiz
     pvlib.iam.physical
     """
-    if _min_scipy():
+    source = _get_model(source_name)
+    target = _get_model(target_name)
 
-        source = _get_model(source_name)
-        target = _get_model(target_name)
+    aoi = np.linspace(0, 90, 100)
+    _check_params(source_name, source_params)
+    source_iam = source(aoi, **source_params)
 
-        aoi = np.linspace(0, 90, 100)
-        _check_params(source_name, source_params)
-        source_iam = source(aoi, **source_params)
-
-        if target_name == "physical":
-            # we can do some special set-up to improve the fit when the
-            # target model is physical
-            if source_name == "ashrae":
-                residual_function, guess, bounds = \
-                    _ashrae_to_physical(aoi, source_iam, weight, fix_n,
-                                        source_params['b'])
-            elif source_name == "martin_ruiz":
-                residual_function, guess, bounds = \
-                    _martin_ruiz_to_physical(aoi, source_iam, weight,
-                                             source_params['a_r'])
-
-        else:
-            # otherwise, target model is ashrae or martin_ruiz, and scipy
-            # does fine without any special set-up
-            bounds = [(1e-04, 1)]
-            guess = [1e-03]
-
-            def residual_function(target_param):
-                return _residual(aoi, source_iam, target, target_param, weight)
-
-        optimize_result = _minimize(residual_function, guess, bounds,
-                                    xtol=xtol)
-
-        return _process_return(target_name, optimize_result)
+    if target_name == "physical":
+        # we can do some special set-up to improve the fit when the
+        # target model is physical
+        if source_name == "ashrae":
+            residual_function, guess, bounds = \
+                _ashrae_to_physical(aoi, source_iam, weight, fix_n,
+                                    source_params['b'])
+        elif source_name == "martin_ruiz":
+            residual_function, guess, bounds = \
+                _martin_ruiz_to_physical(aoi, source_iam, weight,
+                                         source_params['a_r'])
 
     else:
-        return {}
+        # otherwise, target model is ashrae or martin_ruiz, and scipy
+        # does fine without any special set-up
+        bounds = [(1e-04, 1)]
+        guess = [1e-03]
+
+        def residual_function(target_param):
+            return _residual(aoi, source_iam, target, target_param, weight)
+
+    optimize_result = _minimize(residual_function, guess, bounds,
+                                xtol=xtol)
+
+    return _process_return(target_name, optimize_result)
 
 
 def fit(measured_aoi, measured_iam, model_name, weight=_sin_weight, xtol=None):
@@ -1295,30 +1278,27 @@ def fit(measured_aoi, measured_iam, model_name, weight=_sin_weight, xtol=None):
     pvlib.iam.martin_ruiz
     pvlib.iam.physical
     """
-    if _min_scipy():
+    target = _get_model(model_name)
 
-        target = _get_model(model_name)
+    if model_name == "physical":
+        bounds = [(0, 0.08), (1, 2)]
+        guess = [0.002, 1+1e-08]
 
[…]
