+++ docs/examples/reflections/plot_fit_iam_models.py
@@ -70,7 +70,7 @@ plt.show()
 #
 
 # Define a custom weight function.  The weight function must take ``aoi``
-# as it's argument and return a vector of the same length as ``aoi``.
+# as its argument and return a vector of the same length as ``aoi``.
 def weight_function(aoi):
     return cosd(aoi)
 
