### 2023-12-01 0cf1ed4
Add whatsnew entry
