+++ pvlib/pvsystem.py
@@ -7,11 +7,13 @@ from collections import OrderedDict
 import functools
 import io
 import itertools
+import numbers
 import os
 import inspect
 from urllib.request import urlopen
 import numpy as np
 from scipy import constants
+import scipy.optimize
 import pandas as pd
 from dataclasses import dataclass
 from abc import ABC, abstractmethod
@@ -3022,3 +3024,58 @@ def combine_loss_factors(index, *losses, fill_method='ffill'):
         combined_factor *= (1 - loss)
 
     return 1 - combined_factor
+
+
+def _negative_total_power(current, *args):
+    """
+    Compute negative of total power generated by devices in series at specified current.
+    """
+    return -np.sum(current * v_from_i(current, *args))
+
+
+def max_power_point_mismatched(
+        photocurrent, saturation_current, resistance_series, resistance_shunt, nNsVth,
+):
+    """Compute maximum power info for (possibly) mismatched set of devices in series."""
+    if all([isinstance(input, numbers.Number) for input in (
+       photocurrent, saturation_current, resistance_series, resistance_shunt, nNsVth 
+    )]):
+        # There cannot be mismatch, so fall back to non-mismatch function.
+        return {
+            **max_power_point(
+                photocurrent,
+                saturation_current,
+                resistance_series,
+                resistance_shunt,
+                nNsVth
+            )
+        }
+
+    i_mp_0 = max_power_point(
+        np.mean(photocurrent),
+        np.mean(saturation_current),
+        np.mean(resistance_series),
+        np.mean(resistance_shunt),
+        np.mean(nNsVth),
+    )["i_mp"]
+    args = photocurrent, saturation_current, resistance_series, resistance_shunt, nNsVth
+    sol = scipy.optimize.minimize(_negative_total_power, i_mp_0, args)
+
+    if sol.success:
+        i_mp = sol.x[0]
+        v_mp = v_from_i(
+            i_mp,
+            photocurrent,
+            saturation_current,
+            resistance_series,
+            resistance_shunt,
+            nNsVth,
+        )
+
+        return {
+            "i_mp": i_mp,
+            "v_mp": v_mp,
+            "p_mp": i_mp * v_mp,
+        }
+
+    raise RuntimeError(f"unsuccessful solution: {sol}")
