+++ pvlib/pvsystem.py
@@ -3028,18 +3028,34 @@ def combine_loss_factors(index, *losses, fill_method='ffill'):
 
 def _negative_total_power(current, *args):
     """
-    Compute negative of total power generated by devices in series at specified current.
+    Compute negative of total power generated by devices in series at
+    specified current.
     """
     return -np.sum(current * v_from_i(current, *args))
 
 
 def max_power_point_mismatched(
-        photocurrent, saturation_current, resistance_series, resistance_shunt, nNsVth,
+    photocurrent,
+    saturation_current,
+    resistance_series,
+    resistance_shunt,
+    nNsVth,
 ):
-    """Compute maximum power info for (possibly) mismatched set of devices in series."""
-    if all([isinstance(input, numbers.Number) for input in (
-       photocurrent, saturation_current, resistance_series, resistance_shunt, nNsVth 
-    )]):
+    """
+    Compute maximum power info for (possibly) mismatched set of devices in
+    series.
+    """
+    if all(
+        [
+            isinstance(input, numbers.Number) for input in (
+                photocurrent,
+                saturation_current,
+                resistance_series,
+                resistance_shunt,
+                nNsVth,
+            )
+        ]
+    ):
         # There cannot be mismatch, so fall back to non-mismatch function.
         return {
             **max_power_point(
@@ -3058,7 +3074,13 @@ def max_power_point_mismatched(
         np.mean(resistance_shunt),
         np.mean(nNsVth),
     )["i_mp"]
-    args = photocurrent, saturation_current, resistance_series, resistance_shunt, nNsVth
+    args = (
+        photocurrent,
+        saturation_current,
+        resistance_series,
+        resistance_shunt,
+        nNsVth,
+    )
     sol = scipy.optimize.minimize(_negative_total_power, i_mp_0, args)
 
     if sol.success:
