### 2023-12-01 e881ce7
Improve solver, streamline,  and add tests
