+++ pvlib/tests/test_pvsystem.py
@@ -2565,6 +2565,16 @@ def test_Array_temperature_missing_parameters(model, keys):
             "T": 25.0,
         },
         {
+            "photocurrent": 6.2,
+            "saturation_current": 1.0e-8,
+            "n": 1.1,
+            "resistance_series": 0.0001,
+            "resistance_shunt": 5000.0,
+            "Ns": 60,
+            "T": 25.0,
+            "i_mp_ic": 5.8,
+        },
+        {
             "photocurrent": np.array([5.8, 6.2]),
             "saturation_current": 1.0e-8,
             "n": 1.1,
@@ -2572,11 +2582,41 @@ def test_Array_temperature_missing_parameters(model, keys):
             "resistance_shunt": 5000.0,
             "Ns": 60,
             "T": 25.0,
-        }
+        },
+        {
+            "photocurrent": np.array([5.8, 6.2]),
+            "saturation_current": 1.0e-8,
+            "n": 1.1,
+            "resistance_series": 0.0001,
+            "resistance_shunt": 5000.0,
+            "Ns": 60,
+            "T": 25.0,
+            "i_mp_ic": None,
+        },
+        {
+            "photocurrent": np.array([5.8, 6.2]),
+            "saturation_current": 1.0e-8,
+            "n": 1.1,
+            "resistance_series": 0.0001,
+            "resistance_shunt": 5000.0,
+            "Ns": 60,
+            "T": 25.0,
+            "i_mp_ic": -1.0e14,
+        },
+        {
+            "photocurrent": np.array([5.8, 6.2]),
+            "saturation_current": 1.0e-8,
+            "n": 1.1,
+            "resistance_series": 0.0001,
+            "resistance_shunt": 5000.0,
+            "Ns": 60,
+            "T": 25.0,
+            "i_mp_ic": 5.6,
+        },
     ]
 )
 def test_max_power_point_mismatched(inputs):
-    """Test max power point computaiton for mismatched devices in series."""
+    """Test max power point computation for mismatched devices in series."""
 
     photocurrent = inputs["photocurrent"]
     saturation_current = inputs["saturation_current"]
@@ -2587,10 +2627,22 @@ def test_max_power_point_mismatched(inputs):
     T_K = scipy.constants.convert_temperature(inputs["T"], "Celsius", "Kelvin")
     nNsVth = inputs["n"] * inputs["Ns"] * k_B_J_per_K * T_K / q_C
 
-    pvsystem.max_power_point_mismatched(
-        photocurrent,
-        saturation_current,
-        resistance_series,
-        resistance_shunt,
-        nNsVth,
-    )
+    if "i_mp_ic" in inputs:
+        result = pvsystem.max_power_point_mismatched(
+            photocurrent,
+            saturation_current,
+            resistance_series,
+            resistance_shunt,
+            nNsVth,
+            i_mp_ic=inputs["i_mp_ic"],
+        )
+    else:
+        result = pvsystem.max_power_point_mismatched(
+            photocurrent,
+            saturation_current,
+            resistance_series,
+            resistance_shunt,
+            nNsVth,
+        )
+
+    print(result)
