### 2023-12-01 cbe1768
Add test for exceptional case
