### 2023-12-01 3ca7556
Catch exceptions in i_mp_ic computation and update docstring
