### 2023-12-01 58e44e0
Test both exceptional cases
