### 2023-12-01 ab0cef2
add note about clearsky_model
