### 2023-12-01 e40af1e
Monkey patch solver to test exception
