+++ pvlib/pvsystem.py
@@ -3028,7 +3028,7 @@ def combine_loss_factors(index, *losses, fill_method='ffill'):
 def _negative_total_power(current, *args):
     """
     Compute negative of total power generated by devices in series at
-    specified current.
+    specified current. Designed for use by scipy.optimize.minimize.
     """
     return -np.sum(current * v_from_i(current, *args))
 
@@ -3043,6 +3043,8 @@ def max_power_point_mismatched(
     i_mp_ic=None,
 ):
     """
+    FIXME Replace this with proper docstring.
+
     Compute maximum power info for (possibly) mismatched set of devices in
     series. When using this serially on time-series data, passing i_mp_ic from
     previous step may speed up computation. Algorithm falls back to automated
