### 2023-12-04 #5994 @kandersolar
Thanks @rkurchin!

Our release procedure for pvlib automatically creates a Zenodo archive based on the GitHub release, meaning the archive title is something like "pvlib/pvlib-python: v0.10.2" and the authors list reflects the much larger number of people that have ever contributed to the repository (see [example](https://zenodo.org/records/8368494)).  Since those won't match this submission, will I need to make a second archive with metadata that does match?
