### 2023-12-04 #5994 @rkurchin
A few small comments/suggestions on the text:
* line 41 – "used via" feels awkward, any reason not to change it to "implemented in" instead? It doesn't seem that this package is a Python wrapper around something in another language...
* line 84 (into start of 85): this is not a sentence 😉
* 86: "conveniences like optionally standardizing the dataset" also feels a bit awkward, what about "conveniences like the option to standardize the dataset" instead?
* 100: missing a period for "etc." in parenthetical
* In the bibliography, the first SAM citation has a weird semicolon in the OSTI listing

This is amongst the smallest number of editorial comments I've ever had to give! 🙌
