+++ paper/paper.bib
@@ -117,7 +117,7 @@
   title = {SAM Photovoltaic Model Technical Reference 2016 Update},
   url = {http://dx.doi.org/10.2172/1429291},
   DOI = {10.2172/1429291},
-  institution = {Office of Scientific and Technical Information (OSTI)},
+  institution = {Office of Scientific {and} Technical Information (OSTI)},
   author = {Gilman,  Paul and DiOrio,  Nicholas A. and Freeman,  Janine M. and Janzou,  Steven and Dobos,  Aron and Ryberg,  David},
   year = {2018},
   month = mar 
