### 2023-12-05 #5994 @kandersolar
Thanks @rkurchin!  The latest draft addresses those points.  Regarding the "line 41 – used via" comment, the idea was to highlight that to use pvlib is to write code, whereas to use most comparable tools is to point and click with a mouse.  Hopefully the revision is clearer!

FYI I expect the next pvlib version to be released in ~2 weeks, at which point (pending confirmation of the question in https://github.com/openjournals/joss-reviews/issues/5994#issuecomment-1838900507) I'll create the zenodo archive for this review.
