+++ pvlib/tools.py
@@ -11,7 +11,7 @@ import warnings
 
 def cosd(angle):
     """
-    Cosine with angle input in degrees
+    Cosine with angle input in degrees.
 
     Parameters
     ----------
@@ -23,14 +23,13 @@ def cosd(angle):
     result : float or array-like
         Cosine of the angle
     """
-
     res = np.cos(np.radians(angle))
     return res
 
 
 def sind(angle):
     """
-    Sine with angle input in degrees
+    Sine with angle input in degrees.
 
     Parameters
     ----------
@@ -42,14 +41,13 @@ def sind(angle):
     result : float
         Sin of the angle
     """
-
     res = np.sin(np.radians(angle))
     return res
 
 
 def tand(angle):
     """
-    Tan with angle input in degrees
+    Tangent with angle input in degrees.
 
     Parameters
     ----------
@@ -61,14 +59,13 @@ def tand(angle):
     result : float
         Tan of the angle
     """
-
     res = np.tan(np.radians(angle))
     return res
 
 
 def asind(number):
     """
-    Inverse Sine returning an angle in degrees
+    Inverse Sine returning an angle in degrees.
 
     Parameters
     ----------
@@ -80,14 +77,13 @@ def asind(number):
     result : float
         arcsin result
     """
-
     res = np.degrees(np.arcsin(number))
     return res
 
 
 def acosd(number):
     """
-    Inverse Cosine returning an angle in degrees
+    Inverse Cosine returning an angle in degrees.
 
     Parameters
     ----------
@@ -99,8 +95,25 @@ def acosd(number):
     result : float
         arccos result
     """
-
     res = np.degrees(np.arccos(number))
+    return res
+
+
+def atand(angle):
+    """
+    Inverse Tangent returning an angle in degrees.
+
+    Parameters
+    ----------
+    number : float
+        Input number
+
+    Returns
+    -------
+    result : float
+        arccos result
+    """
+    res = np.tan(np.radians(angle))
     return res
 
 
