### 2023-12-07 #5994 @kyleniemeyer
If there is a Zenodo archive that matches the version reviewed for JOSS, then that would be fine to use.
