### 2023-12-07 cbf971b
Update docs/examples/irradiance-transposition/plot_rtranpose_limitations.py
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
