### 2023-12-11 35d679c
Update pvlib/tools.py
Co-authored-by: Kevin Anderson <kevin.anderso@gmail.com>
