### 2023-12-11 a40f7fd
Apply suggestions from code review
Thanks for these improvements @adriesse

Co-authored-by: Anton Driesse <anton.driesse@pvperformancelabs.com>
