+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -10,7 +10,7 @@ Enhancements
 * Added the continuous Perez-Driesse transposition model.
   :py:func:`pvlib.irradiance.perez_driesse` (:issue:`1841`, :pull:`1876`)
 * Added a reverse transposition algorithm using the Perez-Driesse model.
-  :py:func:`pvlib.irradiance.rtranspose_driesse_2023`
+  :py:func:`pvlib.irradiance.ghi_from_poa_driesse_2023`
   (:issue:`1901`, :pull:`1907`)
 * :py:func:`pvlib.bifacial.infinite_sheds.get_irradiance` and
   :py:func:`pvlib.bifacial.infinite_sheds.get_irradiance_poa` now include
