### 2023-12-12 f130248
Add doc string changes from adriesse
Co-authored-by: Anton Driesse <anton.driesse@pvperformancelabs.com>
