### 2023-12-12 #5994 @kandersolar
Thank you for the guidance!  We're anticipating to create that archive sometime next week.

I have another question, this one regarding JOSS's markup capabilities.  We would like to somehow indicate the significance of the authorship ordering in the paper.  The best markup I have come up with so far is to add a footnote to one author's name, as demonstrated in this exploratory commit: https://github.com/kandersolar/pvlib-python/commit/43c7353d33648bec3a5f2760a8d2057e063b2e32 and [rendered PDF](https://github.com/pvlib/pvlib-python/files/12766345/paper-order-footnote.pdf).

Ideally I think the explanation would be somewhere at the top of the first page (under the affiliation list, perhaps).  Is there some way to accomplish that in JOSS's rendering system?  Or is there an alternative method that would be preferred by JOSS?
