### 2023-12-12 300aedc
Docstring cleanup: "default None", doi (#1828)
* Align reference

* Update irradiance.py

Exclude .*=| from negative lookahead

^(\s{8}|\s{4})(?!try|else|doi|DOI|Warning|Access|Requests|Note)(\w*): (?!.*:)(.+)$

Reason: just this line :full_moon_with_face:
Remaining matches are type annotations in modelchain.py and pvsystem.py

* Remove parenthesis in types

F: (?<spaces>\s{8}|\s{4})(?<name>\w*)\s?: (?<type>.*) \(optional, default=(?<default>.*)\)$
R:

* Remove none in type definition #1574

1st-F: (?<spaces> {8}| {4})(?<name>\w*) : (?<types>.*), default None$
2nd-F: (?<spaces> {8}| {4})(?<nam
[…]
