### 2023-12-12 c18a004
Implement reverse transposition using Perez-Driesse forward transposition (#1907)
* Add reverse transposition function and two helpers.

* Add missing import.

* Add to docs under transposition until a better place is found/made.

* Minor doc string fixes.

* Add full_output option similar to newton().

* First example for reverse transposition.

* Second example for reverse transposition.

* Some improvements to the two examples.

* Placate flake8.

* Add tests for reverse transpostion.

* Refine examples and fix test.

* Update whatsnew.

* Refine examples.

* Try to get rid of matplotlib wa
[…]
