### 2023-12-12 6cc599d
increase min scipy to 1.5; update comments
