+++ ci/requirements-py3.12.yml
@@ -7,7 +7,7 @@ dependencies:
     - cython
     - ephem
     - h5py
-    # - numba  # not available for 3.12 as of 2023-10-10
+    # - numba  # not available for 3.12 as of 2023-12-12
     - numpy >= 1.16.0
     - pandas >= 0.25.0
     - pip
@@ -21,8 +21,8 @@ dependencies:
     - python=3.12
     - pytz
     - requests
-    - scipy >= 1.4.0
+    - scipy >= 1.5.0
     - statsmodels
     # - pip:
-        # - nrel-pysam>=2.0  # not available for 3.12 as of 2023-10-10
+        # - nrel-pysam>=2.0  # not available for 3.12 as of 2023-12-12
         # - solarfactors  # required shapely<2 isn't available for 3.12
