### 2023-12-12 #5994 @rkurchin
Good question, and one I've never dealt with before! I think what you've done looks good, but I'll again double-check with @openjournals/pe-eics in case there is a standard approach to this that I'm just not aware of.
