### 2023-12-12 #5994 @kyleniemeyer
@kandersolar @rkurchin I think using a footnote as you did is the best way of handling this. There is a built-in option to specify authors who contributed equally (`equal-contrib: true`) in the metadata, but not something like this.
