### 2023-12-12 f430a74
Fix CAMS message error handler (#1905)
* Fix CAMS message error handler - Issue#1799

* Fix CAMS message error handler - Issue#1799

* fix Python Flake8 Linter error

* Fix associated UT and add contribution in WhatsNews

* Fix line code length in test_sodapro.py

* Fix typo in contributor ghuser

* Update geographical coverage description

* correction in maximum longitude available for CAMS Radiation

* revert of last change

* Add doc string changes from adriesse

Co-authored-by: Anton Driesse <anton.driesse@pvperformancelabs.com>

* Flake8 correction

---------

Co-authored-by: Adam R. Jen
[…]
