+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -18,6 +18,8 @@ Enhancements
 
 Bug fixes
 ~~~~~~~~~
+* Fixed CAMS error message handler in
+  :py:func:`pvlib.iotools.get_cams` (:issue:`1799`, :pull:`1905`)
 * Fix mapping of the dew point column to ``temp_dew`` when ``map_variables``
   is True in :py:func:`pvlib.iotools.get_psm3`. (:pull:`1920`)
 
@@ -45,7 +47,8 @@ Contributors
 * Miguel Sánchez de León Peque (:ghuser:`Peque`)
 * Will Hobbs (:ghuser:`williamhobbs`)
 * Anton Driesse (:ghuser:`adriesse`)
+* Gilles Fischer (:ghuser:`GillesFischerV`)
+* Adam R. Jensen (:ghusuer:`AdamRJensen`)
 * :ghuser:`matsuobasho`
 * Harry Jack (:ghuser:`harry-solcast`)
-* Adam R. Jensen (:ghuser:`AdamRJensen`)
 * Kevin Anderson (:ghuser:`kandersolar`)
