### 2023-12-12 304fbb5
Add python 3.12 to CI (#1886)
* Create requirements-py3.12.yml

* add 3.12 to GH Actions test workflows

* whatsnew

* nix solarfactors

* fix yaml syntax

* increase min scipy to 1.5; update comments
