+++ ci/requirements-py3.12.yml
@@ -1,0 +1,28 @@
+name: test_env
+channels:
+    - defaults
+    - conda-forge
+dependencies:
+    - coveralls
+    - cython
+    - ephem
+    - h5py
+    # - numba  # not available for 3.12 as of 2023-12-12
+    - numpy >= 1.16.0
+    - pandas >= 0.25.0
+    - pip
+    - pytest
+    - pytest-cov
+    - pytest-mock
+    - requests-mock
+    - pytest-timeout
+    - pytest-rerunfailures
+    - pytest-remotedata
+    - python=3.12
+    - pytz
+    - requests
+    - scipy >= 1.5.0
+    - statsmodels
+    # - pip:
+        # - nrel-pysam>=2.0  # not available for 3.12 as of 2023-12-12
+        # - solarfactors  # required shapely<2 isn't available for 3.12
