### 2023-12-13 c0fd312
kandersolar feedback (#3)
* addressing feedback from Kandersolar
