### 2023-12-13 be27d85
Review (#4)
added hack for ISO periods in Pandas 0.25 and clearsky parameter maps
