### 2023-12-13 ae06f5f
pass clearsky_model to get_clearsky
