+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -17,6 +17,8 @@ Bug fixes
 ~~~~~~~~~
 * Fix mapping of the dew point column to ``temp_dew`` when ``map_variables``
   is True in :py:func:`pvlib.iotools.get_psm3`. (:pull:`1920`)
+* Fix :py:class:`pvlib.modelchain.ModelChain` to use attribute `clearsky_model`
+  (:pull:`1924`)
 
 Testing
 ~~~~~~~
