### 2023-12-14 0b6a1ee
install pytest-remotedata from conda-forge
The version in default channel is old.  Version in conda-forge is up to date and contains a fix for this deprecation warning:

DeprecationWarning: distutils Version classes are deprecated. Use packaging.version instead.
