+++ pvlib/iotools/bsrn.py
@@ -321,7 +321,7 @@ def parse_bsrn(fbuf, logical_records=('0100',)):
         LR_0100.columns = BSRN_LR0100_COLUMNS
         # Set datetime index
         LR_0100.index = (start_date+pd.to_timedelta(LR_0100['day']-1, unit='d')
-                         + pd.to_timedelta(LR_0100['minute'], unit='T'))
+                         + pd.to_timedelta(LR_0100['minute'], unit='minutes'))
         # Drop empty, minute, and day columns
         LR_0100 = LR_0100.drop(columns=['empty', 'day', 'minute'])
         dfs.append(LR_0100)
@@ -335,7 +335,7 @@ def parse_bsrn(fbuf, logical_records=('0100',)):
                               colspecs=BSRN_LR0300_COL_SPECS,
                               names=BSRN_LR0300_COLUMNS)
         LR_0300.index = (start_date+pd.to_timedelta(LR_0300['day']-1, unit='d')
-                         + pd.to_timedelta(LR_0300['minute'], unit='T'))
+                         + pd.to_timedelta(LR_0300['minute'], unit='minutes'))
         LR_0300 = LR_0300.drop(columns=['day', 'minute']).astype(float)
         dfs.append(LR_0300)
 
@@ -352,7 +352,7 @@ def parse_bsrn(fbuf, logical_records=('0100',)):
         LR_0500 = LR_0500.reindex(sorted(LR_0500.columns), axis='columns')
         LR_0500.columns = BSRN_LR0500_COLUMNS
         LR_0500.index = (start_date+pd.to_timedelta(LR_0500['day']-1, unit='d')
-                         + pd.to_timedelta(LR_0500['minute'], unit='T'))
+                         + pd.to_timedelta(LR_0500['minute'], unit='minutes'))
         LR_0500 = LR_0500.drop(columns=['empty', 'day', 'minute'])
         dfs.append(LR_0500)
 
