### 2023-12-14 95b5564
use pandas bfill()/ffill() instead of fillna(method='...')
Fixes:

FutureWarning: DataFrame.fillna with 'method' is deprecated and will raise in a future version. Use obj.ffill() or obj.bfill() instead.
