+++ pvlib/ivtools/sdm.py
@@ -680,7 +680,7 @@ def _fit_desoto_sandia_diode(ee, voc, vth, tc, specs, const):
     y = voc - specs['beta_voc'] * (tc - const['T0'])
     new_x = sm.add_constant(x)
     res = sm.RLM(y, new_x).fit()
-    return res.params[1]
+    return np.array(res.params)[1]
 
 
 def _initial_iv_params(ivcurves, ee, voc, isc, rsh, nnsvth):
