### 2023-12-14 ae84817
Fix and document clearsky_model for ModelChain (#1924)
* add note about clearsky_model

* whatsnew

* pass clearsky_model to get_clearsky

* add bug fix note

* test for correct argument

* test for correct argument

* test for correct argument
