### 2023-12-15 37ddf59
Delete racking_model check in infer_temperature_model of model_chain
