### 2023-12-18 f9c8888
change internal linspace to one degree intervals
