### 2023-12-18 12ba8ee
Add functions to fit and convert IAM models (#1827)
* Adding functions, docstrings draft

* Updating imports, adding docstrings

* Updated rst file

* Make linter edits

* Docstring experiment

* Added tests

* More linter edits

* Docstrings and linter edits

* Docstrings and linter

* LINTER

* Docstrings edit

* Added more tests

* Annihilate spaces

* Spacing

* Changed default weight function

* Silence numpy warning

* Updating tests to work with new default

* Forgot a comment

* Return dict contains scalars now, instead of arrays

* Adding option to not fix n

* Adding straggler tests

[…]
