+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -15,6 +15,8 @@ Enhancements
 * :py:func:`pvlib.bifacial.infinite_sheds.get_irradiance` and
   :py:func:`pvlib.bifacial.infinite_sheds.get_irradiance_poa` now include
   shaded fraction in returned variables. (:pull:`1871`)
+* Added :py:func:`~pvlib.iam.convert` and :py:func:`~pvlib.iam.fit` that
+  convert between IAM models, and that fit an IAM model to data. (:issue:`1824`, :pull:`1827`)
 
 Bug fixes
 ~~~~~~~~~
@@ -50,6 +52,8 @@ Contributors
 * Miguel Sánchez de León Peque (:ghuser:`Peque`)
 * Will Hobbs (:ghuser:`williamhobbs`)
 * Anton Driesse (:ghuser:`adriesse`)
+* Abigail Jones (:ghuser:`ajonesr`)
+* Cliff Hansen (:ghuser:`cwhanse`)
 * Gilles Fischer (:ghuser:`GillesFischerV`)
 * Adam R. Jensen (:ghusuer:`AdamRJensen`)
 * :ghuser:`matsuobasho`
