+++ docs/sphinx/source/reference/pv_modeling/iam.rst
@@ -17,3 +17,5 @@ Incident angle modifiers
    iam.marion_integrate
    iam.schlick
    iam.schlick_diffuse
+   iam.convert
+   iam.fit
