### 2023-12-18 fa0c606
Grammatical typo correction in ValueError in null temp_model_params check
