### 2023-12-19 f2baf74
Update pvlib/iotools/solcast.py
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
