### 2023-12-19 25448ce
Update pvlib/iotools/solcast.py
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
