### 2023-12-19 a9dac8f
Update pvlib/iotools/solcast.py
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
