### 2023-12-19 abdee5f
Update pvlib/iotools/solcast.py
Co-authored-by: Adam R. Jensen <39184289+AdamRJensen@users.noreply.github.com>
