+++ pvlib/iotools/__init__.py
@@ -27,3 +27,5 @@ from pvlib.iotools.acis import get_acis_nrcc  # noqa: F401
 from pvlib.iotools.acis import get_acis_mpe  # noqa: F401
 from pvlib.iotools.acis import get_acis_station_data  # noqa: F401
 from pvlib.iotools.acis import get_acis_available_stations  # noqa: F401
+from pvlib.iotools.solaranywhere import get_solaranywhere  # noqa: F401
+from pvlib.iotools.solaranywhere import read_solaranywhere  # noqa: F401
