### 2023-12-19 40ed834
Address code review by kandersolar
