### 2023-12-19 593d8ac
added test for _get_solcast
