### 2023-12-19 8057393
Address several numpy/pandas deprecation warnings; clean up test output (#1930)
* install pytest-remotedata from conda-forge

The version in default channel is old.  Version in conda-forge is up to date and contains a fix for this deprecation warning:

DeprecationWarning: distutils Version classes are deprecated. Use packaging.version instead.

* nix `unit='T'` in pd.to_timedelta

Use `unit='minutes'` instead of `unit='T'` ('T' was deprecated in pandas 2.1).  Fixes:

FutureWarning: Unit 'T' is deprecated and will be removed in a future version.

* Fix StringIO error regex for python 3.12

* us
[…]
