+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -19,7 +19,7 @@ Enhancements
   convert between IAM models, and that fit an IAM model to data. (:issue:`1824`, :pull:`1827`)
 * Add :py:func:`pvlib.iotools.read_solaranywhere` and
   :py:func:`pvlib.iotools.get_solaranywhere` for reading and retrieving
-  SolarAnywhere solar irradiance data. (:pull:`1497`, :discussion:`1310`)
+  SolarAnywhere solar irradiance data. (:pull:`1497`, :discuss:`1310`)
 
 Bug fixes
 ~~~~~~~~~
