### 2023-12-19 0bcb74d
Update solaranywhere documentation
