### 2023-12-19 2ddd01f
Add additional solaranywhere tests
