### 2023-12-19 fb7d41c
Update filestructure of get_srml
