+++ pvlib/tests/iotools/test_srml.py
@@ -1,6 +1,7 @@
 from numpy import isnan
 import pandas as pd
 import pytest
+import os
 
 from pvlib.iotools import srml
 from ..conftest import (DATA_DIR, RERUNS, RERUNS_DELAY, assert_index_equal,
@@ -9,16 +10,17 @@ from pvlib._deprecation import pvlibDeprecationWarning
 
 srml_testfile = DATA_DIR / 'SRML-day-EUPO1801.txt'
 
+BASE_URL = 'http://is-solardata01.uoregon.edu/Step3B_Original_Format/'
+
 
 def test_read_srml():
     srml.read_srml(srml_testfile)
 
 
-@pytest.mark.skip(reason="SRML server is undergoing maintenance as of 12-2023")
 @pytest.mark.remote_data
 @pytest.mark.flaky(reruns=RERUNS, reruns_delay=RERUNS_DELAY)
 def test_read_srml_remote():
-    srml.read_srml('http://solardat.uoregon.edu/download/Archive/EUPO1801.txt')
+    srml.read_srml(os.path.join(BASE_URL, 'EUO/EUO_2018/EUPO1801.txt'))
 
 
 def test_read_srml_columns_exist():
@@ -47,11 +49,10 @@ def test_read_srml_nans_exist():
     assert data['dni_0_flag'].iloc[1119] == 99
 
 
-@pytest.mark.skip(reason="SRML server is undergoing maintenance as of 12-2023")
 @pytest.mark.parametrize('url,year,month', [
-    ('http://solardat.uoregon.edu/download/Archive/EUPO1801.txt',
+    (os.path.join(BASE_URL, 'EUO/EUO_2018/EUPO1801.txt'),
      2018, 1),
-    ('http://solardat.uoregon.edu/download/Archive/EUPO1612.txt',
+    (os.path.join(BASE_URL, 'EUO/EUO_2016/EUPO1612.txt'),
      2016, 12),
 ])
 @pytest.mark.remote_data
@@ -78,78 +79,21 @@ def test__map_columns(column, expected):
     assert srml._map_columns(column) == expected
 
 
-@pytest.mark.skip(reason="SRML server is undergoing maintenance as of 12-2023")
 @pytest.mark.remote_data
 @pytest.mark.flaky(reruns=RERUNS, reruns_delay=RERUNS_DELAY)
 def test_get_srml():
-    url = 'http://solardat.uoregon.edu/download/Archive/EUPO1801.txt'
+    url = os.path.join(BASE_URL, 'EUO/EUO_2018/EUPO1801.txt')
     file_data = srml.read_srml(url)
-    requested, _ = srml.get_srml(station='EU', start='2018-01-01',
+    requested, _ = srml.get_srml(station='EUO', start='2018-01-01',
                                  end='2018-01-31')
     assert_frame_equal(file_data, requested)
 
 
-@pytest.mark.skip(reason="SRML server is undergoing maintenance as of 12-2023")
-@fail_on_pvlib_version('0.11')
-@pytest.mark.remote_data
-@pytest.mark.flaky(reruns=RERUNS, reruns_delay=RERUNS_DELAY)
-def test_read_srml_month_from_solardat():
-    url = 'http://solardat.uoregon.edu/download/Archive/EUPO1801.txt'
-    file_data = srml.read_srml(url)
-    with pytest.warns(pvlibDeprecationWarning, match='get_srml instead'):
-        requested = srml.read_srml_month_from_solardat('EU', 2018, 1)
-    assert file_data.equals(requested)
-
-
-@pytest.mark.skip(reason="SRML server is undergoing maintenance as of 12-2023")
-@fail_on_pvlib_version('0.11')
[…]
