### 2023-12-19 d029684
Add ValueError 2-letter test
