### 2023-12-19 736b0cc
Implement review changes from kandersolar
