+++ pvlib/tests/iotools/test_solcast.py
@@ -1,12 +1,12 @@
 import pandas as pd
-from pvlib.iotools import solcast
+import pvlib
 import pytest
 
 
 @pytest.mark.parametrize("endpoint,function,params,json_response", [
     (
         "live/radiation_and_weather",
-        solcast.get_solcast_live,
+        pvlib.iotools.get_solcast_live,
         dict(
             api_key="1234",
             latitude=-33.856784,
@@ -38,7 +38,7 @@ def test_get_solcast_live(
 
     pd.testing.assert_frame_equal(
         function(**params)[0],
-        solcast._solcast2pvlib(
+        pvlib.iotools.solcast._solcast2pvlib(
             pd.DataFrame.from_dict(
                 json_response[list(json_response.keys())[0]])
         )
@@ -48,7 +48,7 @@ def test_get_solcast_live(
 @pytest.mark.parametrize("endpoint,function,params,json_response", [
     (
         "tmy/radiation_and_weather",
-        solcast.get_solcast_tmy,
+        pvlib.iotools.get_solcast_tmy,
         dict(
             api_key="1234",
             latitude=-33.856784,
@@ -80,7 +80,7 @@ def test_get_solcast_tmy(
 
     pd.testing.assert_frame_equal(
         function(**params)[0],
-        solcast._solcast2pvlib(
+        pvlib.iotools.solcast._solcast2pvlib(
             pd.DataFrame.from_dict(
                 json_response[list(json_response.keys())[0]])
         )
@@ -116,14 +116,14 @@ def test_get_solcast_tmy(
     )
 ])
 def test_solcast2pvlib(in_df, out_df):
-    df = solcast._solcast2pvlib(in_df)
+    df = pvlib.iotools.solcast._solcast2pvlib(in_df)
     pd.testing.assert_frame_equal(df.astype(float), out_df.astype(float))
 
 
 @pytest.mark.parametrize("endpoint,function,params,json_response", [
     (
         "historic/radiation_and_weather",
-        solcast.get_solcast_historic,
+        pvlib.iotools.get_solcast_historic,
         dict(
             api_key="1234",
             latitude=-33.856784,
@@ -157,7 +157,7 @@ def test_get_solcast_historic(
 
     pd.testing.assert_frame_equal(
         function(**params)[0],
-        solcast._solcast2pvlib(
+        pvlib.iotools.solcast._solcast2pvlib(
             pd.DataFrame.from_dict(
                 json_response[list(json_response.keys())[0]]
             )
@@ -168,7 +168,7 @@ def test_get_solcast_historic(
 @pytest.mark.parametrize("endpoint,function,params,json_response", [
     (
         "forecast/radiation_and_weather",
-        solcast.get_solcast_forecast,
+        pvlib.iotools.get_solcast_forecast,
         dict(
             api_key="1234",
             latitude=-33.856784,
@@ -204,7 +204,7 @@ def test_get_solcast_forecast(
 
     pd.testing.assert_frame_equal(
         function(**params)[0],
-        solcast._solcast2pvlib(
+        pvlib.iotools.solcast._solcast2pvlib(
             pd.DataFrame.from_dict(
                 json_response[list(json_response.keys())[0]]
             )
