### 2023-12-19 a009d05
Update test_solaranywhere.py
