+++ pvlib/tests/iotools/test_solaranywhere.py
@@ -67,8 +67,8 @@ def test_read_solaranywhere_high_resolution(high_resolution_index):
     assert 'WindSpeedObservationType' in data.columns
     assert 'Particulate Matter 10 (µg/m3)' in data.columns
     # Check that data is parsed correctly
-    assert data.loc['2021-01-01 07:00-0500', 'Albedo'] == 0.6
-    assert data.loc['2021-01-01 07:00-0500', 'WindSpeed (m/s)'] == 0
+    assert data.loc['2021-01-01 07:00:00-05:00', 'Albedo'] == 0.6
+    assert data.loc['2021-01-01 07:00:00-05:00', 'WindSpeed (m/s)'] == 0
     # Assert that the index is parsed correctly
     pd.testing.assert_index_equal(data.index, high_resolution_index)
 
