### 2023-12-19 2f4ada3
Updated error message when temperature_model_parameters is null after inference to be more informative to user
