### 2023-12-20 26942dc
feat: add additional test coverage (#5)
Co-authored-by: Hugh Cutcher <hugh@solcast.com.au>
