### 2023-12-20 13a5bc4
Add test for raising value error when instantiating array without required args to infer temperature_model_parameters
