### 2023-12-20 8107130
Correct if condition in checking temperature_model_parameters after inference
