### 2023-12-20 da4110f
add functionality to load Solcast API data to `iotools` (#1875)
* prototype (#1)

* first iteration

* dynamic period

* docstring

* feedback

* linting

* Update pvlib/iotools/solcast.py

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

* Update pvlib/iotools/solcast.py

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

* Update pvlib/iotools/solcast.py

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

* midpoint docstring

* flak8 formatting

* PR 1875 (#2)

* kandersolar feedback (#3)

* addressing feedback from Kandersolar

* Review (#4)

added hack for ISO periods in Pandas 0.25 and
[…]
