+++ docs/sphinx/source/reference/iotools.rst
@@ -46,6 +46,10 @@ of sources and file formats relevant to solar energy modeling.
    iotools.get_acis_station_data
    iotools.get_acis_available_stations
    iotools.read_panond
+   iotools.get_solcast_tmy
+   iotools.get_solcast_historic
+   iotools.get_solcast_forecast
+   iotools.get_solcast_live
 
 
 A :py:class:`~pvlib.location.Location` object may be created from metadata
