+++ docs/sphinx/source/whatsnew/v0.10.3.rst
@@ -15,6 +15,9 @@ Enhancements
 * :py:func:`pvlib.bifacial.infinite_sheds.get_irradiance` and
   :py:func:`pvlib.bifacial.infinite_sheds.get_irradiance_poa` now include
   shaded fraction in returned variables. (:pull:`1871`)
+* Added :py:func:`~pvlib.iotools.get_solcast_tmy`, :py:func:`~pvlib.iotools.get_solcast_historic`,
+  :py:func:`~pvlib.iotools.get_solcast_forecast` and :py:func:`~pvlib.iotools.get_solcast_live` to
+  read data from the Solcast API. (:issue:`1313`, :pull:`1875`)
 * Added :py:func:`~pvlib.iam.convert` and :py:func:`~pvlib.iam.fit` that
   convert between IAM models, and that fit an IAM model to data. (:issue:`1824`, :pull:`1827`)
 
@@ -59,4 +62,5 @@ Contributors
 * :ghuser:`matsuobasho`
 * Harry Jack (:ghuser:`harry-solcast`)
 * Kevin Anderson (:ghuser:`kandersolar`)
+* Lorenzo Riches (:ghuser:`lorenzo-solcast`)
 * Cliff Hansen (:ghuser:`cwhanse`)
