### 2023-12-20 62ba7d9
Apply suggestions from code review
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
