+++ pvlib/iotools/solaranywhere.py
@@ -53,7 +53,7 @@ def get_solaranywhere(latitude, longitude, api_key, start=None, end=None,
     """Retrieve historical irradiance time series data from SolarAnywhere.
 
     The SolarAnywhere API is described in [1]_ and [2]_. A detailed list of
-    available options for the input parameters can be found in [3]_.
+    API options can be found in [3]_.
 
     Parameters
     ----------
@@ -77,12 +77,12 @@ def get_solaranywhere(latitude, longitude, api_key, start=None, end=None,
         'SolarAnywhere3_2' (see [3]_ for a full list of options).
     time_resolution: {60, 30, 15, 5}, default: 60
         Time resolution in minutes. For TMY data, time resolution has to be 60
-        min. (hourly).
+        minutes (hourly).
     spatial_resolution: {0.1, 0.01, 0.005}, default: 0.01
         Spatial resolution in degrees.
     true_dynamics: bool, default: False
         Whether to apply SolarAnywhere TrueDynamics statistical processing.
-        Only available for the 5-min time resolution.
+        Only available for the 5-minute time resolution.
     probability_of_exceedance: int, optional
         Probability of exceedance in the range of 1 to 99. Only relevant when
         requesting probability of exceedance (POE) time series. [%]
@@ -96,7 +96,7 @@ def get_solaranywhere(latitude, longitude, api_key, start=None, end=None,
         Base url of SolarAnywhere API.
     map_variables: bool, default: True
         When true, renames columns of the DataFrame to pvlib variable names
-        where applicable. See variable :const:`VARIABLE_MAP`.
+        where applicable. See :const:`VARIABLE_MAP`.
     timeout: float, default: 300
         Time in seconds to wait for requested data to become available.
 
@@ -225,7 +225,7 @@ def read_solaranywhere(filename, map_variables=True, encoding='iso-8859-1'):
 
     The SolarAnywhere file format and variables are described in [1]_. Note,
     the SolarAnywhere file format resembles the TMY3 file format but contains
-    additional variables and meatadata.
+    additional variables and metadata.
 
     Parameters
     ----------
@@ -233,7 +233,7 @@ def read_solaranywhere(filename, map_variables=True, encoding='iso-8859-1'):
         Filename
     map_variables: bool, default: True
         When true, renames columns of the DataFrame to pvlib variable names
-        where applicable. See variable :const:`VARIABLE_MAP`.
+        where applicable. See :const:`VARIABLE_MAP`.
     encoding : str, default : 'iso-8859-1'
         Encoding of the file. For SolarAnywhere TMY3 files the 'iso-8859-1'
         encoding is recommended due to the usage of special characters.
