### 2023-12-20 6db7fa5
Switch to isinstance
