### 2023-12-20 006719a
Update docs/sphinx/source/whatsnew/v0.10.3.rst
Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
