### 2023-12-20 ac70b14
Add functions to read and retrieve SolarAnywhere irradiance data (#1497)
* Add cams.get_cams_radiation function

* Revert "Add cams.get_cams_radiation function"

This reverts commit d7deb80cdc5d1b63de5b2865a0c5cf24d4655fc1.

* Add get, read, parse solaranywhere functions

* Add whatsnew

* Updates to get_solaranywhere

* Minor doc updates

* Updated default values & add POE

* Properly raise start/end ValueError

* Add api_key to pytest-remote-data.yml

* Add test coverage

* Set encoding to iso-8859-1

* Remove solaranywhere_api_key

* Remove parse_solaranywhere

I am removing the parse_solar
[…]
