### 2023-12-20 1601760
Finalize 0.10.3 (#1934)
* clean up 0.10.3 whatsnew

* Create v0.10.4.rst

* Update docs/sphinx/source/whatsnew/v0.10.3.rst

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>

---------

Co-authored-by: Cliff Hansen <cwhanse@sandia.gov>
