### 2023-12-20 #5994 @kandersolar
Great, I've updated the manuscript to include that footnote.  Thank you for the guidance.

@rkurchin we have just released pvlib v0.10.3, with corresponding archive: https://doi.org/10.5281/zenodo.10412885

If anything else is needed from me for this submission, please let me know :)
