### 2023-12-20 2bafd68
Added some additional details to the error message to make it informative for user to diagnose
