### 2023-12-21 5f2abdf
Add raises section in docstring of Array to describe error
