+++ docs/sphinx/source/user_guide/pvsystem.rst
@@ -62,10 +62,15 @@ that describe a PV system's modules and inverter are stored in
 
 .. ipython:: python
 
+    from pvlib.temperature import TEMPERATURE_MODEL_PARAMETERS
+
+    temp_params = TEMPERATURE_MODEL_PARAMETERS["pvsyst"]["freestanding"]
+
     module_parameters = {'pdc0': 5000, 'gamma_pdc': -0.004}
     inverter_parameters = {'pdc0': 5000, 'eta_inv_nom': 0.96}
     system = pvsystem.PVSystem(inverter_parameters=inverter_parameters,
-                               module_parameters=module_parameters)
+                               module_parameters=module_parameters,
+                               temperature_model_parameters = temp_params)
     print(system.inverter_parameters)
 
 
@@ -125,7 +130,8 @@ passed to `PVSystem.module_parameters`:
     module_parameters = {'pdc0': 5000, 'gamma_pdc': -0.004}
     inverter_parameters = {'pdc0': 5000, 'eta_inv_nom': 0.96}
     system = pvsystem.PVSystem(module_parameters=module_parameters,
-                               inverter_parameters=inverter_parameters)
+                               inverter_parameters=inverter_parameters,
+                               temperature_model_parameters=temp_params)
     print(system.arrays[0].module_parameters)
     print(system.inverter_parameters)
 
@@ -140,16 +146,16 @@ provided for each array, and the arrays are provided to
 .. ipython:: python
 
     module_parameters = {'pdc0': 5000, 'gamma_pdc': -0.004}
-    mount = pvsystem.FixedMount(surface_tilt=20, surface_azimuth=180)
-    array_one = pvsystem.Array(mount=mount, module_parameters=module_parameters)
-    array_two = pvsystem.Array(mount=mount, module_parameters=module_parameters)
+    mount = pvsystem.FixedMount(surface_tilt=20, surface_azimuth=180, racking_model='close_mount')
+    array_one = pvsystem.Array(mount=mount, module_parameters=module_parameters, module_type='glass_glass')
+    array_two = pvsystem.Array(mount=mount, module_parameters=module_parameters, module_type='glass_glass')
     system_two_arrays = pvsystem.PVSystem(arrays=[array_one, array_two],
                                           inverter_parameters=inverter_parameters)
     print([array.module_parameters for array in system_two_arrays.arrays])
     print(system_two_arrays.inverter_parameters)
 
 
-The :py:class:`~pvlib.pvsystem.Array` class includes those 
+The :py:class:`~pvlib.pvsystem.Array` class includes those
 :py:class:`~pvlib.pvsystem.PVSystem` attributes that may vary from array
 to array. These attributes include
 `module_parameters`, `temperature_model_parameters`, `modules_per_string`,
@@ -171,7 +177,7 @@ PVSystem attributes
 -------------------
 
 Here we review the most commonly used PVSystem and Array attributes.
-Please see the :py:class:`~pvlib.pvsystem.PVSystem` and 
+Please see the :py:class:`~pvlib.pvsystem.PVSystem` and
 :py:class:`~pvlib.pvsystem.Array` class documentation for a
 comprehensive list of attributes.
 
@@ -182,14 +188,27 @@ Tilt and azimuth
 The first parameters which describe the DC part of a PV system are the tilt
 and azimuth of the modules. In the case of a PV system with a single array,
 these parameters can be specified using the `PVSystem.surface_tilt` and
-`PVSystem.surface_azimuth` attributes.  This will automatically create
+`PVSystem.surface_azimuth` attributes.
+
+Temperature model parameters
[…]
