### 2023-12-21 #5994 @rkurchin
Noting that the suggested pytest DOI is a book chapter about it and doesn't seem to be "officially" associated with the project. I can't find evidence of an official DOI, and the `CITATION` file on the repo doesn't have one, so I think this is probably fine as-is.
