### 2023-12-22 196be6a
Update instances of pvsystem instantiation in test_pvsystem tests to reflect new required argument
