+++ pvlib/tests/test_pvsystem.py
@@ -30,22 +30,25 @@ from pvlib.tests.test_singlediode import get_pvsyst_fs_495
     ('physical', {'K': 4, 'L': 0.002, 'n': 1.526}),
     ('martin_ruiz', {'a_r': 0.16}),
 ])
-def test_PVSystem_get_iam(mocker, iam_model, model_params):
+def test_PVSystem_get_iam(mocker, iam_model, model_params, temp_model_params):
     m = mocker.spy(_iam, iam_model)
-    system = pvsystem.PVSystem(module_parameters=model_params)
+    system = pvsystem.PVSystem(module_parameters=model_params,
+                               temperature_model_parameters=temp_model_params)
     thetas = 1
     iam = system.get_iam(thetas, iam_model=iam_model)
     m.assert_called_with(thetas, **model_params)
     assert iam < 1.
 
 
-def test_PVSystem_multi_array_get_iam():
+def test_PVSystem_multi_array_get_iam(temp_model_params):
     model_params = {'b': 0.05}
     system = pvsystem.PVSystem(
         arrays=[pvsystem.Array(mount=pvsystem.FixedMount(0, 180),
-                               module_parameters=model_params),
+                               module_parameters=model_params,
+                               temperature_model_parameters=temp_model_params),
                 pvsystem.Array(mount=pvsystem.FixedMount(0, 180),
-                               module_parameters=model_params)]
+                               module_parameters=model_params,
+                               temperature_model_parameters=temp_model_params)]
     )
     iam = system.get_iam((1, 5), iam_model='ashrae')
     assert len(iam) == 2
@@ -55,8 +58,9 @@ def test_PVSystem_multi_array_get_iam():
         system.get_iam((1,), iam_model='ashrae')
 
 
-def test_PVSystem_get_iam_sapm(sapm_module_params, mocker):
-    system = pvsystem.PVSystem(module_parameters=sapm_module_params)
+def test_PVSystem_get_iam_sapm(sapm_module_params, mocker, temp_model_params):
+    system = pvsystem.PVSystem(module_parameters=sapm_module_params,
+                               temperature_model_parameters=temp_model_params)
     mocker.spy(_iam, 'sapm')
     aoi = 0
     out = system.get_iam(aoi, 'sapm')
@@ -64,9 +68,10 @@ def test_PVSystem_get_iam_sapm(sapm_module_params, mocker):
     assert_allclose(out, 1.0, atol=0.01)
 
 
-def test_PVSystem_get_iam_interp(mocker):
+def test_PVSystem_get_iam_interp(mocker, temp_model_params):
     interp_module_params = {'iam_ref': (1., 0.8), 'theta_ref': (0., 80.)}
-    system = pvsystem.PVSystem(module_parameters=interp_module_params)
+    system = pvsystem.PVSystem(module_parameters=interp_module_params,
+                               temperature_model_parameters=temp_model_params)
     spy = mocker.spy(_iam, 'interp')
     aoi = ((0., 40., 80.),)
     expected = (1., 0.9, 0.8)
@@ -98,8 +103,9 @@ def test__normalize_sam_product_names():
     assert list(norm_names) == NORM_NAMES
 
 
-def test_PVSystem_get_iam_invalid(sapm_module_params, mocker):
-    system = pvsystem.PVSystem(module_parameters=sapm_module_params)
+def test_PVSystem_get_iam_invalid(sapm_module_params, temp_model_params, mocker):
+    system = pvsystem.PVSystem(module_parameters=sapm_module_params,
+                               temperature_model_parameters=temp_model_params)
     with pytest.raises(ValueError):
         system.get_iam(45, iam_model='not_a_model')
 
@@ -226,9 +232,10 @@ def test_sapm(sapm_module_params):
[…]
