### 2023-12-22 #5994 @kyleniemeyer
Hi @kandersolar, I'm doing some final checks before publishing.

I noticed that the author list of the Zenodo archive doesn't match that of the paper; the paper authors show up at the beginning of the Zenodo author list, but not in the same order and there are some others at the top of the list as well.

Normally we expect these to match, although I see that the paper author list is ordered based on commits. Can you comment on this?

In the paper, I see a few minor issues:
- Is the first reference to Figure 1 supposed to be part of the preceding paragraph, or the beginning of a new one? If the latter, I think `\autoref{fig:functions-comparison}` needs to start on the preceding line, and if the latter, you need a line break before it.
- Can you remove "e.g" before the reference to Gilman et al. 2019? I think it is implied with a citation like that, and then we can eliminate the nested parentheses.
- Similarly, I'd love to get rid of the nested parentheses in that long parenthetical on the second paragraph of the statement of need (they are a pet peeve of mine...). Could you split that long sentence and parenthetical up a bit?

For example, maybe something like
```
Other PV system modeling tools like SAM [@Gilman2018], PVsyst [@Mermoud1994],
SolarFarmer [@Mikofski2018], PlantPredict [@Passow2017], and
CASSYS [@Pai2016]—to name a few software tools with comparable breadth of
modeling capability—organize the user interface around pre-built modeling workflows.
Instead, pvlib python makes the individual "building blocks" of PV system performance
models accessible to the user.
```
