### 2023-12-22 #5994 @kandersolar
> In the paper, I see a few minor issues:

Thanks for these improvements!  I think the latest proof addresses these points.

> Can you comment on this?

Sure!  Regarding the Zenodo archive: we have an integration set up so that tagging a new release on GitHub automatically triggers the creation of a corresponding Zenodo archive.  I don't 100% understand how Zenodo determines author lists, but the list on the archive we're using for this submission seems to be the top 30 contributors (out of 100+ total) by total commit count over the project's full history (2013-present).  Compare with what GitHub reports for that list here: https://github.com/pvlib/pvlib-python/graphs/contributors

For this JOSS submission, the authors list reflects the set of "core developers" (people with repo edit privileges and/or sustained contribution to project maintenance and direction) over just the period since the project's last JOSS paper (2018-present), with ordering determined by commit count over that same period.

So the differences come from what level of contribution to the project (core developer vs all code contributors) and what period of contribution (since the 2018 JOSS paper vs all time) are being considered.
