### 2023-12-22 #5994 @kandersolar
@kyleniemeyer, one small note: the metadata displayed on the JOSS paper page is showing stray brackets after "Anderson", I guess because of that footnote I included:

![image](https://github.com/openjournals/joss-reviews/assets/57452607/cafc695c-b108-4880-9f50-0cda0bb26dac)
