### 2024-01-19 #5994 @arfon
This is due to the way the author has written a footnote. @tarleb doe you recall what the preferred way of accomplishing this is? It doesn't seem to be in our docs.
