### 2024-01-19 #5994 @tarleb
I think the preferred way is to use inline footnote syntax, e.g. `^[Note on author order]`. Let me check to make sure.
