### 2024-01-19 #5994 @tarleb
I can't check if the above syntax works with the website, but it works fine with the PDF generator.
