### 2024-02-13 #5994 @kandersolar
@arfon Just to confirm, there is no action needed from me to fix this footnote issue after the fact, correct?
