### 2024-04-03 #5994 @AdamRJensen
@tarleb did you manage to find a fix for the footnote issue? It still renders incorrectly on the [JOSS website](https://joss.theoj.org/papers/10.21105/joss.05994).
